% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Autofit}
\alias{Autofit}
\title{Autofit Object}
\usage{
Autofit(fontScale = NULL, autofitType = NULL, lineSpacingReduction = NULL)
}
\arguments{
\item{fontScale}{The font scale applied to the shape}

\item{autofitType}{The autofit type of the shape}

\item{lineSpacingReduction}{The line spacing reduction applied to the shape}
}
\value{
Autofit object
}
\description{
Autofit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The autofit properties of a Shape.
}
\concept{Autofit functions}
