% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{AutoText}
\alias{AutoText}
\title{AutoText Object}
\usage{
AutoText(type = NULL, content = NULL, style = NULL)
}
\arguments{
\item{type}{(string)
The type of this auto text.
\itemize{
\item \code{TYPE_UNSPECIFIED}: An unspecified autotext type.
\item \code{SLIDE_NUMBER}: Type for autotext that represents the current slide number.
}}

\item{content}{(string)
The rendered content of this auto text, if available.}

\item{style}{(\link{TextStyle})
The styling applied to this auto text.}
}
\value{
AutoText object
}
\description{
AutoText Object
}
\details{
A TextElement kind that represents auto text.
}
