% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{BatchUpdatePresentationRequest}
\alias{BatchUpdatePresentationRequest}
\title{BatchUpdatePresentationRequest Object}
\usage{
BatchUpdatePresentationRequest(requests = NULL, writeControl = NULL)
}
\arguments{
\item{requests}{(list of \link{Request} objects)
A list of updates to apply to the presentation.}

\item{writeControl}{(\link{WriteControl})
Provides control over how write requests are executed.}
}
\value{
BatchUpdatePresentationRequest object
}
\description{
BatchUpdatePresentationRequest Object
}
\details{
Request message for PresentationsService.BatchUpdatePresentation.
}
