% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{BatchUpdatePresentationResponse}
\alias{BatchUpdatePresentationResponse}
\title{BatchUpdatePresentationResponse Object}
\usage{
BatchUpdatePresentationResponse(
  presentationId = NULL,
  replies = NULL,
  writeControl = NULL
)
}
\arguments{
\item{presentationId}{(string)
The presentation the updates were applied to.}

\item{replies}{(list of \link{Response} objects)
The reply of the updates. This maps 1:1 with the updates, although replies to some requests may be empty.}

\item{writeControl}{(\link{WriteControl})
The updated write control after applying the request.}
}
\value{
BatchUpdatePresentationResponse object
}
\description{
BatchUpdatePresentationResponse Object
}
\details{
Response message from a batch update.
}
