% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Bullet}
\alias{Bullet}
\title{Bullet Object}
\usage{
Bullet(listId = NULL, nestingLevel = NULL, glyph = NULL, bulletStyle = NULL)
}
\arguments{
\item{listId}{(string)
The ID of the list this paragraph belongs to.}

\item{nestingLevel}{(integer)
The nesting level of this paragraph in the list.}

\item{glyph}{(string)
The rendered bullet glyph for this paragraph.}

\item{bulletStyle}{(\link{TextStyle})
The paragraph specific text style applied to this bullet.}
}
\value{
Bullet object
}
\description{
Bullet Object
}
\details{
Describes the bullet of a paragraph.
}
