% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.R
\name{choose_slides}
\alias{choose_slides}
\title{Choose Slides presentation}
\usage{
choose_slides()
}
\value{
A presentation id
}
\description{
Opens a webpage for a user to authenticate with Google and select a presentation.
This presentation is then authorised for use with SlidesTools.
}
