% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ColorScheme}
\alias{ColorScheme}
\title{ColorScheme Object}
\usage{
ColorScheme(colors = NULL)
}
\arguments{
\item{colors}{(list of \link{ThemeColorPair} objects)
The ThemeColorType and corresponding concrete color pairs.}
}
\value{
ColorScheme object
}
\description{
ColorScheme Object
}
\details{
The palette of predefined colors for a page.
}
