% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ColorStop}
\alias{ColorStop}
\title{ColorStop Object}
\usage{
ColorStop(position = NULL, color = NULL, alpha = NULL)
}
\arguments{
\item{position}{The relative position of the color stop in the gradient band measured in percentage}

\item{color}{The color of the gradient stop}

\item{alpha}{The alpha value of this color in the gradient band}
}
\value{
ColorStop object
}
\description{
ColorStop Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A color and position in a gradient band.
}
\concept{ColorStop functions}
