% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateImageResponse}
\alias{CreateImageResponse}
\title{CreateImageResponse Object}
\usage{
CreateImageResponse(objectId = NULL)
}
\arguments{
\item{objectId}{(string)
The object ID of the created image.}
}
\value{
CreateImageResponse object
}
\description{
CreateImageResponse Object
}
\details{
The result of creating an image.
}
