% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateLineResponse}
\alias{CreateLineResponse}
\title{CreateLineResponse Object}
\usage{
CreateLineResponse(objectId = NULL)
}
\arguments{
\item{objectId}{(string)
The object ID of the created line.}
}
\value{
CreateLineResponse object
}
\description{
CreateLineResponse Object
}
\details{
The result of creating a line.
}
