% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateParagraphBulletsRequest}
\alias{CreateParagraphBulletsRequest}
\title{CreateParagraphBulletsRequest Object}
\usage{
CreateParagraphBulletsRequest(
  objectId = NULL,
  cellLocation = NULL,
  textRange = NULL,
  bulletPreset = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID of the shape or table containing the text to add bullets to.}

\item{cellLocation}{(\link{TableCellLocation})
The optional table cell location if the text to be modified is in a table cell. If present, the object_id must refer to a table.}

\item{textRange}{(\link{Range})
The range of text to apply the bullet presets to, based on TextElement indexes.}

\item{bulletPreset}{(string)
The kinds of bullet glyphs to be used. Defaults to the \code{BULLET_DISC_CIRCLE_SQUARE} preset.}
}
\value{
CreateParagraphBulletsRequest object
}
\description{
CreateParagraphBulletsRequest Object
}
\details{
Creates bullets for all of the paragraphs that overlap with the given text index range. The nesting level of each paragraph will be determined by counting leading tabs in front of each paragraph. To avoid excess space between the bullet and the corresponding paragraph, these leading tabs are removed by this request. This may change the indices of parts of the text. If the paragraph immediately before paragraphs being updated is in a list with a matching preset, the paragraphs being updated are added to that preceding list.
}
