% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateShapeRequest}
\alias{CreateShapeRequest}
\title{CreateShapeRequest Object}
\usage{
CreateShapeRequest(objectId = NULL, elementProperties = NULL, shapeType = NULL)
}
\arguments{
\item{objectId}{(string)
A user-supplied object ID. If you specify an ID, it must be unique among all pages and page elements in the presentation. The ID must start with an alphanumeric character or an underscore (matches regex \verb{[a-zA-Z0-9_]}); remaining characters may include those as well as a hyphen or colon (matches regex \verb{[a-zA-Z0-9_-:]}). The length of the ID must not be less than 5 or greater than 50. If empty, a unique identifier will be generated.}

\item{elementProperties}{(\link{PageElementProperties})
The element properties for the shape.}

\item{shapeType}{(string)
The shape type.}
}
\value{
CreateShapeRequest object
}
\description{
CreateShapeRequest Object
}
\details{
Creates a new shape.
}
