% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateSheetsChartRequest}
\alias{CreateSheetsChartRequest}
\title{CreateSheetsChartRequest Object}
\usage{
CreateSheetsChartRequest(
  spreadsheetId = NULL,
  linkingMode = NULL,
  elementProperties = NULL,
  chartId = NULL,
  objectId = NULL
)
}
\arguments{
\item{spreadsheetId}{The ID of the Google Sheets spreadsheet that contains the chart}

\item{linkingMode}{The mode with which the chart is linked to the source spreadsheet}

\item{elementProperties}{The element properties for the chart}

\item{chartId}{The ID of the specific chart in the Google Sheets spreadsheet}

\item{objectId}{A user-supplied object ID}
}
\value{
CreateSheetsChartRequest object
}
\description{
CreateSheetsChartRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creates an embedded Google Sheets chart. NOTE: Chart creation requires at least one of the spreadsheets.readonly, spreadsheets, drive.readonly, drive.file, or drive OAuth scopes.
}
\concept{CreateSheetsChartRequest functions}
