% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateSlideRequest}
\alias{CreateSlideRequest}
\title{CreateSlideRequest Object}
\usage{
CreateSlideRequest(
  objectId = NULL,
  insertionIndex = NULL,
  slideLayoutReference = NULL,
  placeholderIdMappings = NULL
)
}
\arguments{
\item{objectId}{A user-supplied object ID}

\item{insertionIndex}{The optional zero-based index indicating where to insert the slides}

\item{slideLayoutReference}{Layout reference of the slide to be inserted, based on the *current master*, which is one of the following: - The master of the previous slide index}

\item{placeholderIdMappings}{An optional list of object ID mappings from the placeholder(s) on the layout to the placeholders that are created on the slide from the specified layout}
}
\value{
CreateSlideRequest object
}
\description{
CreateSlideRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creates a slide.
}
\concept{CreateSlideRequest functions}
