% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateSlideResponse}
\alias{CreateSlideResponse}
\title{CreateSlideResponse Object}
\usage{
CreateSlideResponse(objectId = NULL)
}
\arguments{
\item{objectId}{The object ID of the created slide}
}
\value{
CreateSlideResponse object
}
\description{
CreateSlideResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of creating a slide.
}
\concept{CreateSlideResponse functions}
