% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateTableRequest}
\alias{CreateTableRequest}
\title{CreateTableRequest Object}
\usage{
CreateTableRequest(
  objectId = NULL,
  elementProperties = NULL,
  rows = NULL,
  columns = NULL
)
}
\arguments{
\item{objectId}{(string)
A user-supplied object ID. If you specify an ID, it must be unique among all pages and page elements in the presentation. The ID must start with an alphanumeric character or an underscore (matches regex \verb{[a-zA-Z0-9_]}); remaining characters may include those as well as a hyphen or colon (matches regex \verb{[a-zA-Z0-9_-:]}). The length of the ID must not be less than 5 or greater than 50. If you don't specify an ID, a unique one is generated.}

\item{elementProperties}{(\link{PageElementProperties})
The element properties for the table. The table will be created at the provided size, subject to a minimum size. If no size is provided, the table will be automatically sized. Table transforms must have a scale of 1 and no shear components. If no transform is provided, the table will be centered on the page.}

\item{rows}{(integer)
Number of rows in the table.}

\item{columns}{(integer)
Number of columns in the table.}
}
\value{
CreateTableRequest object
}
\description{
CreateTableRequest Object
}
\details{
Creates a new table.
}
