% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateTableResponse}
\alias{CreateTableResponse}
\title{CreateTableResponse Object}
\usage{
CreateTableResponse(objectId = NULL)
}
\arguments{
\item{objectId}{(string)
The object ID of the created table.}
}
\value{
CreateTableResponse object
}
\description{
CreateTableResponse Object
}
\details{
The result of creating a table.
}
