% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateVideoRequest}
\alias{CreateVideoRequest}
\title{CreateVideoRequest Object}
\usage{
CreateVideoRequest(
  objectId = NULL,
  elementProperties = NULL,
  source = NULL,
  id = NULL
)
}
\arguments{
\item{objectId}{A user-supplied object ID}

\item{elementProperties}{The element properties for the video}

\item{source}{The video source}

\item{id}{The video source's unique identifier for this video}
}
\value{
CreateVideoRequest object
}
\description{
CreateVideoRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creates a video. NOTE: Creating a video from Google Drive requires that the requesting app have at least one of the drive, drive.readonly, or drive.file OAuth scopes.
}
\concept{CreateVideoRequest functions}
