% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateVideoResponse}
\alias{CreateVideoResponse}
\title{CreateVideoResponse Object}
\usage{
CreateVideoResponse(objectId = NULL)
}
\arguments{
\item{objectId}{(string)
The object ID of the created video.}
}
\value{
CreateVideoResponse object
}
\description{
CreateVideoResponse Object
}
\details{
The result of creating a video.
}
