% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{DeleteObjectRequest}
\alias{DeleteObjectRequest}
\title{DeleteObjectRequest Object}
\usage{
DeleteObjectRequest(objectId = NULL)
}
\arguments{
\item{objectId}{(string)
The object ID of the page or page element to delete. If after a delete operation a group contains only 1 or no page elements, the group is also deleted. If a placeholder is deleted on a layout, any empty inheriting placeholders are also deleted.}
}
\value{
DeleteObjectRequest object
}
\description{
DeleteObjectRequest Object
}
\details{
Deletes an object, either pages or page elements, from the presentation.
}
