% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{DeleteParagraphBulletsRequest}
\alias{DeleteParagraphBulletsRequest}
\title{DeleteParagraphBulletsRequest Object}
\usage{
DeleteParagraphBulletsRequest(
  objectId = NULL,
  cellLocation = NULL,
  textRange = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID of the shape or table containing the text to delete bullets from.}

\item{cellLocation}{(\link{TableCellLocation})
The optional table cell location if the text to be modified is in a table cell. If present, the object_id must refer to a table.}

\item{textRange}{(\link{Range})
The range of text to delete bullets from, based on TextElement indexes.}
}
\value{
DeleteParagraphBulletsRequest object
}
\description{
DeleteParagraphBulletsRequest Object
}
\details{
Deletes bullets from all of the paragraphs that overlap with the given text index range. The nesting level of each paragraph will be visually preserved by adding indent to the start of the corresponding paragraph.
}
