% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{DeleteTableColumnRequest}
\alias{DeleteTableColumnRequest}
\title{DeleteTableColumnRequest Object}
\usage{
DeleteTableColumnRequest(tableObjectId = NULL, cellLocation = NULL)
}
\arguments{
\item{tableObjectId}{(string)
The table to delete columns from.}

\item{cellLocation}{(\link{TableCellLocation})
The reference table cell location from which a column will be deleted. The column this cell spans will be deleted. If this is a merged cell, multiple columns will be deleted. If no columns remain in the table after this deletion, the whole table is deleted.}
}
\value{
DeleteTableColumnRequest object
}
\description{
DeleteTableColumnRequest Object
}
\details{
Deletes a column from a table.
}
