% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{DeleteTableRowRequest}
\alias{DeleteTableRowRequest}
\title{DeleteTableRowRequest Object}
\usage{
DeleteTableRowRequest(tableObjectId = NULL, cellLocation = NULL)
}
\arguments{
\item{tableObjectId}{The table to delete rows from}

\item{cellLocation}{The reference table cell location from which a row will be deleted}
}
\value{
DeleteTableRowRequest object
}
\description{
DeleteTableRowRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deletes a row from a table.
}
\concept{DeleteTableRowRequest functions}
