% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Dimension}
\alias{Dimension}
\title{Dimension Object}
\usage{
Dimension(magnitude = NULL, unit = NULL)
}
\arguments{
\item{magnitude}{(number)
The magnitude.}

\item{unit}{(string)
The units for magnitude.}
}
\value{
Dimension object
}
\description{
Dimension Object
}
\details{
A magnitude in a single direction in the specified units.
}
