% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{DuplicateObjectRequest}
\alias{DuplicateObjectRequest}
\title{DuplicateObjectRequest Object}
\usage{
DuplicateObjectRequest(
  DuplicateObjectRequest.objectIds = NULL,
  objectId = NULL,
  objectIds = NULL
)
}
\arguments{
\item{DuplicateObjectRequest.objectIds}{The \link{DuplicateObjectRequest.objectIds} object or list of objects}

\item{objectId}{The ID of the object to duplicate}

\item{objectIds}{The object being duplicated may contain other objects, for example when duplicating a slide or a group page element}
}
\value{
DuplicateObjectRequest object
}
\description{
DuplicateObjectRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Duplicates a slide or page element. When duplicating a slide, the duplicate slide will be created immediately following the specified slide. When duplicating a page element, the duplicate will be placed on the same page at the same position as the original.
}
\seealso{
Other DuplicateObjectRequest functions: 
\code{\link{DuplicateObjectRequest.objectIds}()}
}
\concept{DuplicateObjectRequest functions}
