% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{DuplicateObjectResponse}
\alias{DuplicateObjectResponse}
\title{DuplicateObjectResponse Object}
\usage{
DuplicateObjectResponse(objectId = NULL)
}
\arguments{
\item{objectId}{(string)
The ID of the new duplicate object.}
}
\value{
DuplicateObjectResponse object
}
\description{
DuplicateObjectResponse Object
}
\details{
The response of duplicating an object.
}
