% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleslides.api-package.R
\docType{package}
\name{googleslides.api-package}
\alias{googleslides.api}
\alias{googleslides.api-package}
\title{googleslides.api: Google Slides API client and tools}
\description{
Create Google Slides presentations from within R with full API support.
}
\author{
\strong{Maintainer}: Isaac Gravestock \email{isaac.gravestock@gmail.com} [compiler]

Other contributors:
\itemize{
  \item Google (Createor of API and documentation) [copyright holder]
}

}
\keyword{internal}
