% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Group}
\alias{Group}
\title{Group Object}
\usage{
Group(children = NULL)
}
\arguments{
\item{children}{(list of \link{PageElement} objects)
The collection of elements in the group. The minimum size of a group is 2.}
}
\value{
Group object
}
\description{
Group Object
}
\details{
A PageElement kind representing a joined collection of PageElements.
}
