% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{GroupObjectsResponse}
\alias{GroupObjectsResponse}
\title{GroupObjectsResponse Object}
\usage{
GroupObjectsResponse(objectId = NULL)
}
\arguments{
\item{objectId}{The object ID of the created group}
}
\value{
GroupObjectsResponse object
}
\description{
GroupObjectsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of grouping objects.
}
\concept{GroupObjectsResponse functions}
