% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsd_auth.R
\name{gsd_deauth}
\alias{gsd_deauth}
\title{Suspend authorization}
\usage{
gsd_deauth()
}
\description{
Put gslideDevice into a de-authorized state. Instead of sending a token,
gslideDevice will send an API key. This can be used to access public
resources for which no Google sign-in is required. This is handy for using
gslideDevice in a non-interactive setting to make requests that do not
require a token. It will prevent the attempt to obtain a token
interactively in the browser. The user can configure their own API key
via \code{\link[=gsd_auth_configure]{gsd_auth_configure()}} and retrieve that key via
\code{\link[=gsd_api_key]{gsd_api_key()}}.
In the absence of a user-configured key, a built-in default key is used.
}
\examples{
\dontrun{
gsd_deauth()
gsd_user()
}
}
\seealso{
Other auth functions: 
\code{\link{gsd_auth}()},
\code{\link{gsd_auth_configure}()}
}
\concept{auth functions}
