% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsd_auth.R
\name{gsd_has_token}
\alias{gsd_has_token}
\title{Is there a token on hand?}
\usage{
gsd_has_token()
}
\value{
Logical.
}
\description{
Reports whether gslideDevice has stored a token, ready for use in downstream
requests.
}
\examples{
gsd_has_token()
}
\seealso{
Other low-level API functions: 
\code{\link{gsd_token}()}
}
\concept{low-level API functions}
