% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsd_auth.R
\name{gsd_token}
\alias{gsd_token}
\title{Produce configured token}
\usage{
gsd_token()
}
\value{
A `request` object (an S3 class provided by [httr][httr::httr]).
}
\description{
For internal use or for those programming around the Slides API.
Returns a token pre-processed with [httr::config()]. Most users
do not need to handle tokens "by hand" or, even if they need some
control, [gsd_auth()] is what they need. If there is no current
token, [gsd_auth()] is called to either load from cache or
initiate OAuth2.0 flow.
If auth has been deactivated via [gsd_deauth()], `gsd_token()`
returns `NULL`.
}
\examples{
\dontrun{
gsd_token()
}
}
\seealso{
Other low-level API functions: 
\code{\link{gsd_has_token}()}
}
\concept{low-level API functions}
