% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Image}
\alias{Image}
\title{Image Object}
\usage{
Image(
  contentUrl = NULL,
  imageProperties = NULL,
  sourceUrl = NULL,
  placeholder = NULL
)
}
\arguments{
\item{contentUrl}{(string)
An URL to an image with a default lifetime of 30 minutes. This URL is tagged with the account of the requester. Anyone with the URL effectively accesses the image as the original requester. Access to the image may be lost if the presentation's sharing settings change.}

\item{imageProperties}{(\link{ImageProperties})
The properties of the image.}

\item{sourceUrl}{(string)
The source URL is the URL used to insert the image. The source URL can be empty.}

\item{placeholder}{(\link{Placeholder})
Placeholders are page elements that inherit from corresponding placeholders on layouts and masters. If set, the image is a placeholder image and any inherited properties can be resolved by looking at the parent placeholder identified by the Placeholder.parent_object_id field.}
}
\value{
Image object
}
\description{
Image Object
}
\details{
A PageElement kind representing an image.
}
