% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ImageProperties}
\alias{ImageProperties}
\title{ImageProperties Object}
\usage{
ImageProperties(
  cropProperties = NULL,
  transparency = NULL,
  brightness = NULL,
  contrast = NULL,
  recolor = NULL,
  outline = NULL,
  shadow = NULL,
  link = NULL
)
}
\arguments{
\item{cropProperties}{(\link{CropProperties})
The crop properties of the image. If not set, the image is not cropped. This property is read-only.}

\item{transparency}{(number)
The transparency effect of the image. The value should be in the interval [0.0, 1.0], where 0 means no effect and 1 means completely transparent. This property is read-only.}

\item{brightness}{(number)
The brightness effect of the image. The value should be in the interval [-1.0, 1.0], where 0 means no effect. This property is read-only.}

\item{contrast}{(number)
The contrast effect of the image. The value should be in the interval [-1.0, 1.0], where 0 means no effect. This property is read-only.}

\item{recolor}{(\link{Recolor})
The recolor effect of the image. If not set, the image is not recolored. This property is read-only.}

\item{outline}{(\link{Outline})
The outline of the image. If not set, the image has no outline.}

\item{shadow}{(\link{Shadow})
The shadow of the image. If not set, the image has no shadow. This property is read-only.}

\item{link}{(\link{Link})
The hyperlink destination of the image. If unset, there is no link.}
}
\value{
ImageProperties object
}
\description{
ImageProperties Object
}
\details{
The properties of the Image.
}
