% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{InsertTableRowsRequest}
\alias{InsertTableRowsRequest}
\title{InsertTableRowsRequest Object}
\usage{
InsertTableRowsRequest(
  tableObjectId = NULL,
  cellLocation = NULL,
  insertBelow = NULL,
  number = NULL
)
}
\arguments{
\item{tableObjectId}{(string)
The table to insert rows into.}

\item{cellLocation}{(\link{TableCellLocation})
The reference table cell location from which rows will be inserted. A new row will be inserted above (or below) the row where the reference cell is. If the reference cell is a merged cell, a new row will be inserted above (or below) the merged cell.}

\item{insertBelow}{(boolean)
Whether to insert new rows below the reference cell location. - \code{True}: insert below the cell. - \code{False}: insert above the cell.}

\item{number}{(integer)
The number of rows to be inserted. Maximum 20 per request.}
}
\value{
InsertTableRowsRequest object
}
\description{
InsertTableRowsRequest Object
}
\details{
Inserts rows into a table.
}
