% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ladder.api-package}
\alias{ladder.api}
\alias{ladder.api-package}
\title{ladder.api: Google Slides API client and tools}
\description{
Create, read and modify Slides presentations with full REST API functionality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://igrave.github.io/SlidesTools/}
}

}
\author{
\strong{Maintainer}: Isaac Gravestock \email{isaac.gravestock@gmail.com} [compiler]

Other contributors:
\itemize{
  \item Google (Createor of API and documentation) [copyright holder]
}

}
\keyword{internal}
