% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_auth.R
\name{ladder_has_token}
\alias{ladder_has_token}
\title{Is there a token on hand?}
\usage{
ladder_has_token()
}
\value{
Logical.
}
\description{
Reports whether ladder.api has stored a token, ready for use in downstream
requests.
}
\examples{
ladder_has_token()
}
\seealso{
Other low-level API functions: 
\code{\link{ladder_token}()}
}
\concept{low-level API functions}
