% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{LayoutPlaceholderIdMapping}
\alias{LayoutPlaceholderIdMapping}
\title{LayoutPlaceholderIdMapping Object}
\usage{
LayoutPlaceholderIdMapping(
  layoutPlaceholder = NULL,
  objectId = NULL,
  layoutPlaceholderObjectId = NULL
)
}
\arguments{
\item{layoutPlaceholder}{The placeholder on a layout that will be applied to a slide}

\item{objectId}{A user-supplied object ID for the placeholder identified above that to be created onto a slide}

\item{layoutPlaceholderObjectId}{The object ID of the placeholder on a layout that will be applied to a slide}
}
\value{
LayoutPlaceholderIdMapping object
}
\description{
LayoutPlaceholderIdMapping Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The user-specified ID mapping for a placeholder that will be created on a slide from a specified layout.
}
\concept{LayoutPlaceholderIdMapping functions}
