% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{LayoutProperties}
\alias{LayoutProperties}
\title{LayoutProperties Object}
\usage{
LayoutProperties(masterObjectId = NULL, name = NULL, displayName = NULL)
}
\arguments{
\item{masterObjectId}{(string)
The object ID of the master that this layout is based on.}

\item{name}{(string)
The name of the layout.}

\item{displayName}{(string)
The human-readable name of the layout.}
}
\value{
LayoutProperties object
}
\description{
LayoutProperties Object
}
\details{
The properties of Page are only relevant for pages with page_type LAYOUT.
}
