% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Line}
\alias{Line}
\title{Line Object}
\usage{
Line(lineProperties = NULL, lineType = NULL, lineCategory = NULL)
}
\arguments{
\item{lineProperties}{The properties of the line}

\item{lineType}{The type of the line}

\item{lineCategory}{The category of the line}
}
\value{
Line object
}
\description{
Line Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A PageElement kind representing a non-connector line, straight connector, curved connector, or bent connector.
}
\concept{Line functions}
