% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{LineConnection}
\alias{LineConnection}
\title{LineConnection Object}
\usage{
LineConnection(connectedObjectId = NULL, connectionSiteIndex = NULL)
}
\arguments{
\item{connectedObjectId}{The object ID of the connected page element}

\item{connectionSiteIndex}{The index of the connection site on the connected page element}
}
\value{
LineConnection object
}
\description{
LineConnection Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties for one end of a Line connection.
}
\concept{LineConnection functions}
