% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{LineFill}
\alias{LineFill}
\title{LineFill Object}
\usage{
LineFill(solidFill = NULL)
}
\arguments{
\item{solidFill}{(\link{SolidFill})
Solid color fill.}
}
\value{
LineFill object
}
\description{
LineFill Object
}
\details{
The fill of the line.
}
