% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{List}
\alias{List}
\title{List Object}
\usage{
List(listId = NULL, nestingLevel = NULL)
}
\arguments{
\item{listId}{(string)
The ID of the list.}

\item{nestingLevel}{(named list of \link{NestingLevel} objects)
A map of nesting levels to the properties of bullets at the associated level. A list has at most nine levels of nesting, so the possible values for the keys of this map are 0 through 8, inclusive.}
}
\value{
List object
}
\description{
List Object
}
\details{
A List describes the look and feel of bullets belonging to paragraphs associated with a list. A paragraph that is part of a list has an implicit reference to that list's ID.
}
