% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{MergeTableCellsRequest}
\alias{MergeTableCellsRequest}
\title{MergeTableCellsRequest Object}
\usage{
MergeTableCellsRequest(objectId = NULL, tableRange = NULL)
}
\arguments{
\item{objectId}{(string)
The object ID of the table.}

\item{tableRange}{(\link{TableRange})
The table range specifying which cells of the table to merge. Any text in the cells being merged will be concatenated and stored in the upper-left ("head") cell of the range. If the range is non-rectangular (which can occur in some cases where the range covers cells that are already merged), a 400 bad request error is returned.}
}
\value{
MergeTableCellsRequest object
}
\description{
MergeTableCellsRequest Object
}
\details{
Merges cells in a Table.
}
