% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{NestingLevel}
\alias{NestingLevel}
\title{NestingLevel Object}
\usage{
NestingLevel(bulletStyle = NULL)
}
\arguments{
\item{bulletStyle}{(\link{TextStyle})
The style of a bullet at this level of nesting.}
}
\value{
NestingLevel object
}
\description{
NestingLevel Object
}
\details{
Contains properties describing the look and feel of a list bullet at a given level of nesting.
}
