% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{OpaqueColor}
\alias{OpaqueColor}
\title{OpaqueColor Object}
\usage{
OpaqueColor(rgbColor = NULL, themeColor = NULL)
}
\arguments{
\item{rgbColor}{(\link{RgbColor})
An opaque RGB color.}

\item{themeColor}{(string)
An opaque theme color.}
}
\value{
OpaqueColor object
}
\description{
OpaqueColor Object
}
\details{
A themeable solid color value.
}
