% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{OptionalColor}
\alias{OptionalColor}
\title{OptionalColor Object}
\usage{
OptionalColor(opaqueColor = NULL)
}
\arguments{
\item{opaqueColor}{(\link{OpaqueColor})
If set, this will be used as an opaque color. If unset, this represents a transparent color.}
}
\value{
OptionalColor object
}
\description{
OptionalColor Object
}
\details{
A color that can either be fully opaque or fully transparent.
}
