% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{OutlineFill}
\alias{OutlineFill}
\title{OutlineFill Object}
\usage{
OutlineFill(solidFill = NULL)
}
\arguments{
\item{solidFill}{(\link{SolidFill})
Solid color fill.}
}
\value{
OutlineFill object
}
\description{
OutlineFill Object
}
\details{
The fill of the outline.
}
