% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{PageBackgroundFill}
\alias{PageBackgroundFill}
\title{PageBackgroundFill Object}
\usage{
PageBackgroundFill(
  solidFill = NULL,
  propertyState = NULL,
  stretchedPictureFill = NULL
)
}
\arguments{
\item{solidFill}{Solid color fill}

\item{propertyState}{The background fill property state}

\item{stretchedPictureFill}{Stretched picture fill}
}
\value{
PageBackgroundFill object
}
\description{
PageBackgroundFill Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The page background fill.
}
\concept{PageBackgroundFill functions}
