% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{PageElement}
\alias{PageElement}
\title{PageElement Object}
\usage{
PageElement(
  image = NULL,
  line = NULL,
  elementGroup = NULL,
  title = NULL,
  transform = NULL,
  shape = NULL,
  video = NULL,
  sheetsChart = NULL,
  objectId = NULL,
  table = NULL,
  description = NULL,
  size = NULL,
  wordArt = NULL
)
}
\arguments{
\item{image}{An image page element}

\item{line}{A line page element}

\item{elementGroup}{A collection of page elements joined as a single unit}

\item{title}{The title of the page element}

\item{transform}{The transform of the page element}

\item{shape}{A generic shape}

\item{video}{A video page element}

\item{sheetsChart}{A linked chart embedded from Google Sheets}

\item{objectId}{The object ID for this page element}

\item{table}{A table page element}

\item{description}{The description of the page element}

\item{size}{The size of the page element}

\item{wordArt}{A word art page element}
}
\value{
PageElement object
}
\description{
PageElement Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A visual element rendered on a page.
}
\concept{PageElement functions}
