% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{PageElementProperties}
\alias{PageElementProperties}
\title{PageElementProperties Object}
\usage{
PageElementProperties(pageObjectId = NULL, size = NULL, transform = NULL)
}
\arguments{
\item{pageObjectId}{The object ID of the page where the element is located}

\item{size}{The size of the element}

\item{transform}{The transform for the element}
}
\value{
PageElementProperties object
}
\description{
PageElementProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Common properties for a page element. Note: When you initially create a PageElement, the API may modify the values of both `size` and `transform`, but the visual size will be unchanged.
}
\concept{PageElementProperties functions}
