% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{PageProperties}
\alias{PageProperties}
\title{PageProperties Object}
\usage{
PageProperties(pageBackgroundFill = NULL, colorScheme = NULL)
}
\arguments{
\item{pageBackgroundFill}{(\link{PageBackgroundFill})
The background fill of the page. If unset, the background fill is inherited from a parent page if it exists. If the page has no parent, then the background fill defaults to the corresponding fill in the Slides editor.}

\item{colorScheme}{(\link{ColorScheme})
The color scheme of the page. If unset, the color scheme is inherited from a parent page. If the page has no parent, the color scheme uses a default Slides color scheme, matching the defaults in the Slides editor. Only the concrete colors of the first 12 ThemeColorTypes are editable. In addition, only the color scheme on \code{Master} pages can be updated. To update the field, a color scheme containing mappings from all the first 12 ThemeColorTypes to their concrete colors must be provided. Colors for the remaining ThemeColorTypes will be ignored.}
}
\value{
PageProperties object
}
\description{
PageProperties Object
}
\details{
The properties of the Page. The page will inherit properties from the parent page. Depending on the page type the hierarchy is defined in either SlideProperties or LayoutProperties.
}
