% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ParagraphMarker}
\alias{ParagraphMarker}
\title{ParagraphMarker Object}
\usage{
ParagraphMarker(bullet = NULL, style = NULL)
}
\arguments{
\item{bullet}{The bullet for this paragraph}

\item{style}{The paragraph's style}
}
\value{
ParagraphMarker object
}
\description{
ParagraphMarker Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TextElement kind that represents the beginning of a new paragraph.
}
\concept{ParagraphMarker functions}
