% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Placeholder}
\alias{Placeholder}
\title{Placeholder Object}
\usage{
Placeholder(type = NULL, index = NULL, parentObjectId = NULL)
}
\arguments{
\item{type}{(string)
The type of the placeholder.}

\item{index}{(integer)
The index of the placeholder. If the same placeholder types are present in the same page, they would have different index values.}

\item{parentObjectId}{(string)
The object ID of this shape's parent placeholder. If unset, the parent placeholder shape does not exist, so the shape does not inherit properties from any other shape.}
}
\value{
Placeholder object
}
\description{
Placeholder Object
}
\details{
The placeholder information that uniquely identifies a placeholder shape.
}
