% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{presentations.create}
\alias{presentations.create}
\title{Creates a blank presentation using the title given in the request.}
\usage{
presentations.create(Presentation)
}
\arguments{
\item{Presentation}{The \link{Presentation} object to pass to this method}
}
\description{
If a \code{presentationId} is provided, it is used as the ID of the new
presentation. Otherwise, a new ID is generated. Other fields in the request,
including any provided content, are ignored. Returns the created presentation.
}
\seealso{
\href{https://developers.google.com/slides/}{Google Documentation}
}
\concept{Presentation functions}
