% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{presentations.get}
\alias{presentations.get}
\title{Gets the latest version of the specified presentation.}
\usage{
presentations.get(presentationId)
}
\arguments{
\item{presentationId}{The ID of the presentation to retrieve}
}
\description{
Gets the latest version of the specified presentation.
}
\seealso{
\href{https://developers.google.com/slides/}{Google Documentation}
}
