% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{presentations.pages.get}
\alias{presentations.pages.get}
\title{Gets the latest version of the specified page in the presentation.}
\usage{
presentations.pages.get(presentationId, pageObjectId)
}
\arguments{
\item{presentationId}{The ID of the presentation to retrieve}

\item{pageObjectId}{The object ID of the page to retrieve}
}
\description{
Gets the latest version of the specified page in the presentation.
}
\seealso{
\href{https://developers.google.com/slides/}{Google Documentation}
}
