% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{presentations.pages.getThumbnail}
\alias{presentations.pages.getThumbnail}
\title{Generates a thumbnail of the latest version of the specified page in the presentation and returns a URL to the thumbnail image. This request counts as an [expensive read request](/slides/limits) for quota purposes.}
\usage{
presentations.pages.getThumbnail(
  presentationId,
  pageObjectId,
  thumbnailProperties.thumbnailSize = NULL,
  thumbnailProperties.mimeType = NULL
)
}
\arguments{
\item{presentationId}{The ID of the presentation to retrieve}

\item{pageObjectId}{The object ID of the page whose thumbnail to retrieve}

\item{thumbnailProperties.thumbnailSize}{The optional thumbnail image size}

\item{thumbnailProperties.mimeType}{The optional mime type of the thumbnail image}
}
\description{
Generates a thumbnail of the latest version of the specified page in the presentation and returns a URL to the thumbnail image. This request counts as an [expensive read request](/slides/limits) for quota purposes.
}
\seealso{
\href{https://developers.google.com/slides/}{Google Documentation}
}
