% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Range}
\alias{Range}
\title{Range Object}
\usage{
Range(startIndex = NULL, endIndex = NULL, type = NULL)
}
\arguments{
\item{startIndex}{(integer)
The optional zero-based index of the beginning of the collection. Required for \code{FIXED_RANGE} and \code{FROM_START_INDEX} ranges.}

\item{endIndex}{(integer)
The optional zero-based index of the end of the collection. Required for \code{FIXED_RANGE} ranges.}

\item{type}{(string)
The type of range.
\itemize{
\item \code{RANGE_TYPE_UNSPECIFIED}: Unspecified range type. This value must not be used.
\item \code{FIXED_RANGE}: A fixed range. Both the \code{start_index} and \code{end_index} must be specified.
\item \code{FROM_START_INDEX}: Starts the range at \code{start_index} and continues until the end of the collection. The \code{end_index} must not be specified.
\item \code{ALL}: Sets the range to be the whole length of the collection. Both the \code{start_index} and the \code{end_index} must not be specified.
}}
}
\value{
Range object
}
\description{
Range Object
}
\details{
Specifies a contiguous range of an indexed collection, such as characters in text.
}
