% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Recolor}
\alias{Recolor}
\title{Recolor Object}
\usage{
Recolor(recolorStops = NULL, name = NULL)
}
\arguments{
\item{recolorStops}{(list of \link{ColorStop} objects)
The recolor effect is represented by a gradient, which is a list of color stops. The colors in the gradient will replace the corresponding colors at the same position in the color palette and apply to the image. This property is read-only.}

\item{name}{(string)
The name of the recolor effect. The name is determined from the \code{recolor_stops} by matching the gradient against the colors in the page's current color scheme. This property is read-only.}
}
\value{
Recolor object
}
\description{
Recolor Object
}
\details{
A recolor effect applied on an image.
}
