% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ReplaceAllShapesWithImageRequest}
\alias{ReplaceAllShapesWithImageRequest}
\title{ReplaceAllShapesWithImageRequest Object}
\usage{
ReplaceAllShapesWithImageRequest(
  imageReplaceMethod = NULL,
  replaceMethod = NULL,
  containsText = NULL,
  imageUrl = NULL,
  pageObjectIds = NULL
)
}
\arguments{
\item{imageReplaceMethod}{The image replace method}

\item{replaceMethod}{The replace method}

\item{containsText}{If set, this request will replace all of the shapes that contain the given text}

\item{imageUrl}{The image URL}

\item{pageObjectIds}{If non-empty, limits the matches to page elements only on the given pages}
}
\value{
ReplaceAllShapesWithImageRequest object
}
\description{
ReplaceAllShapesWithImageRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Replaces all shapes that match the given criteria with the provided image. The images replacing the shapes are rectangular after being inserted into the presentation and do not take on the forms of the shapes.
}
\concept{ReplaceAllShapesWithImageRequest functions}
