% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ReplaceAllShapesWithImageResponse}
\alias{ReplaceAllShapesWithImageResponse}
\title{ReplaceAllShapesWithImageResponse Object}
\usage{
ReplaceAllShapesWithImageResponse(occurrencesChanged = NULL)
}
\arguments{
\item{occurrencesChanged}{(integer)
The number of shapes replaced with images.}
}
\value{
ReplaceAllShapesWithImageResponse object
}
\description{
ReplaceAllShapesWithImageResponse Object
}
\details{
The result of replacing shapes with an image.
}
