% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ReplaceAllShapesWithSheetsChartRequest}
\alias{ReplaceAllShapesWithSheetsChartRequest}
\title{ReplaceAllShapesWithSheetsChartRequest Object}
\usage{
ReplaceAllShapesWithSheetsChartRequest(
  containsText = NULL,
  chartId = NULL,
  linkingMode = NULL,
  pageObjectIds = NULL,
  spreadsheetId = NULL
)
}
\arguments{
\item{containsText}{The criteria that the shapes must match in order to be replaced}

\item{chartId}{The ID of the specific chart in the Google Sheets spreadsheet}

\item{linkingMode}{The mode with which the chart is linked to the source spreadsheet}

\item{pageObjectIds}{If non-empty, limits the matches to page elements only on the given pages}

\item{spreadsheetId}{The ID of the Google Sheets spreadsheet that contains the chart}
}
\value{
ReplaceAllShapesWithSheetsChartRequest object
}
\description{
ReplaceAllShapesWithSheetsChartRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Replaces all shapes that match the given criteria with the provided Google Sheets chart. The chart will be scaled and centered to fit within the bounds of the original shape. NOTE: Replacing shapes with a chart requires at least one of the spreadsheets.readonly, spreadsheets, drive.readonly, or drive OAuth scopes.
}
\concept{ReplaceAllShapesWithSheetsChartRequest functions}
