% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ReplaceAllTextRequest}
\alias{ReplaceAllTextRequest}
\title{ReplaceAllTextRequest Object}
\usage{
ReplaceAllTextRequest(
  replaceText = NULL,
  pageObjectIds = NULL,
  containsText = NULL
)
}
\arguments{
\item{replaceText}{(string)
The text that will replace the matched text.}

\item{pageObjectIds}{(list of \link{string} objects)
If non-empty, limits the matches to page elements only on the given pages. Returns a 400 bad request error if given the page object ID of a notes master, or if a page with that object ID doesn't exist in the presentation.}

\item{containsText}{(\link{SubstringMatchCriteria})
Finds text in a shape matching this substring.}
}
\value{
ReplaceAllTextRequest object
}
\description{
ReplaceAllTextRequest Object
}
\details{
Replaces all instances of text matching a criteria with replace text.
}
