% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ReplaceAllTextResponse}
\alias{ReplaceAllTextResponse}
\title{ReplaceAllTextResponse Object}
\usage{
ReplaceAllTextResponse(occurrencesChanged = NULL)
}
\arguments{
\item{occurrencesChanged}{(integer)
The number of occurrences changed by replacing all text.}
}
\value{
ReplaceAllTextResponse object
}
\description{
ReplaceAllTextResponse Object
}
\details{
The result of replacing text.
}
