% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ReplaceImageRequest}
\alias{ReplaceImageRequest}
\title{ReplaceImageRequest Object}
\usage{
ReplaceImageRequest(
  imageReplaceMethod = NULL,
  imageObjectId = NULL,
  url = NULL
)
}
\arguments{
\item{imageReplaceMethod}{The replacement method}

\item{imageObjectId}{The ID of the existing image that will be replaced}

\item{url}{The image URL}
}
\value{
ReplaceImageRequest object
}
\description{
ReplaceImageRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Replaces an existing image with a new image. Replacing an image removes some image effects from the existing image.
}
\concept{ReplaceImageRequest functions}
