% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Request}
\alias{Request}
\title{Request Object}
\usage{
Request(...)
}
\arguments{
\item{insertTableColumns}{Inserts columns into a table}

\item{deleteTableColumn}{Deletes a column from a table}

\item{updateSlidesPosition}{Updates the position of a set of slides in the presentation}

\item{updateTextStyle}{Updates the styling of text within a Shape or Table}

\item{createSlide}{Creates a new slide}

\item{updateParagraphStyle}{Updates the styling of paragraphs within a Shape or Table}

\item{updateTableRowProperties}{Updates the properties of a Table row}

\item{updateTableBorderProperties}{Updates the properties of the table borders in a Table}

\item{replaceAllShapesWithSheetsChart}{Replaces all shapes matching some criteria with a Google Sheets chart}

\item{updateTableColumnProperties}{Updates the properties of a Table column}

\item{ungroupObjects}{Ungroups objects, such as groups}

\item{replaceImage}{Replaces an existing image with a new image}

\item{deleteTableRow}{Deletes a row from a table}

\item{createParagraphBullets}{Creates bullets for paragraphs}

\item{updateLineProperties}{Updates the properties of a Line}

\item{updateImageProperties}{Updates the properties of an Image}

\item{deleteParagraphBullets}{Deletes bullets from paragraphs}

\item{createSheetsChart}{Creates an embedded Google Sheets chart}

\item{updateVideoProperties}{Updates the properties of a Video}

\item{updateShapeProperties}{Updates the properties of a Shape}

\item{groupObjects}{Groups objects, such as page elements}

\item{mergeTableCells}{Merges cells in a Table}

\item{replaceAllShapesWithImage}{Replaces all shapes matching some criteria with an image}

\item{refreshSheetsChart}{Refreshes a Google Sheets chart}

\item{unmergeTableCells}{Unmerges cells in a Table}

\item{updateSlideProperties}{Updates the properties of a Slide}

\item{updatePageElementTransform}{Updates the transform of a page element}

\item{insertTableRows}{Inserts rows into a table}

\item{deleteText}{Deletes text from a shape or a table cell}

\item{createImage}{Creates an image}

\item{createVideo}{Creates a video}

\item{updatePageElementAltText}{Updates the alt text title and/or description of a page element}

\item{createLine}{Creates a line}

\item{rerouteLine}{Reroutes a line such that it's connected at the two closest connection sites on the connected page elements}

\item{createTable}{Creates a new table}

\item{updateLineCategory}{Updates the category of a line}

\item{updatePageElementsZOrder}{Updates the Z-order of page elements}

\item{updatePageProperties}{Updates the properties of a Page}

\item{createShape}{Creates a new shape}

\item{updateTableCellProperties}{Updates the properties of a TableCell}

\item{insertText}{Inserts text into a shape or table cell}

\item{duplicateObject}{Duplicates a slide or page element}

\item{deleteObject}{Deletes a page or page element from the presentation}

\item{replaceAllText}{Replaces all instances of specified text}
}
\value{
Request object
}
\description{
Request Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single kind of update to apply to a presentation.
}
\concept{Request functions}
