% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{RerouteLineRequest}
\alias{RerouteLineRequest}
\title{RerouteLineRequest Object}
\usage{
RerouteLineRequest(objectId = NULL)
}
\arguments{
\item{objectId}{The object ID of the line to reroute}
}
\value{
RerouteLineRequest object
}
\description{
RerouteLineRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Reroutes a line such that it's connected at the two closest connection sites on the connected page elements.
}
\concept{RerouteLineRequest functions}
