% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Response}
\alias{Response}
\title{Response Object}
\usage{
Response(
  createSlide = NULL,
  createShape = NULL,
  createTable = NULL,
  replaceAllText = NULL,
  createImage = NULL,
  createVideo = NULL,
  createSheetsChart = NULL,
  createLine = NULL,
  replaceAllShapesWithImage = NULL,
  duplicateObject = NULL,
  replaceAllShapesWithSheetsChart = NULL,
  groupObjects = NULL
)
}
\arguments{
\item{createSlide}{(\link{CreateSlideResponse})
The result of creating a slide.}

\item{createShape}{(\link{CreateShapeResponse})
The result of creating a shape.}

\item{createTable}{(\link{CreateTableResponse})
The result of creating a table.}

\item{replaceAllText}{(\link{ReplaceAllTextResponse})
The result of replacing text.}

\item{createImage}{(\link{CreateImageResponse})
The result of creating an image.}

\item{createVideo}{(\link{CreateVideoResponse})
The result of creating a video.}

\item{createSheetsChart}{(\link{CreateSheetsChartResponse})
The result of creating a Google Sheets chart.}

\item{createLine}{(\link{CreateLineResponse})
The result of creating a line.}

\item{replaceAllShapesWithImage}{(\link{ReplaceAllShapesWithImageResponse})
The result of replacing all shapes matching some criteria with an image.}

\item{duplicateObject}{(\link{DuplicateObjectResponse})
The result of duplicating an object.}

\item{replaceAllShapesWithSheetsChart}{(\link{ReplaceAllShapesWithSheetsChartResponse})
The result of replacing all shapes matching some criteria with a Google Sheets chart.}

\item{groupObjects}{(\link{GroupObjectsResponse})
The result of grouping objects.}
}
\value{
Response object
}
\description{
Response Object
}
\details{
A single response from an update.
}
