% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{RgbColor}
\alias{RgbColor}
\title{RgbColor Object}
\usage{
RgbColor(red = NULL, green = NULL, blue = NULL)
}
\arguments{
\item{red}{(number)
The red component of the color, from 0.0 to 1.0.}

\item{green}{(number)
The green component of the color, from 0.0 to 1.0.}

\item{blue}{(number)
The blue component of the color, from 0.0 to 1.0.}
}
\value{
RgbColor object
}
\description{
RgbColor Object
}
\details{
An RGB color.
}
