% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Shape}
\alias{Shape}
\title{Shape Object}
\usage{
Shape(
  shapeType = NULL,
  text = NULL,
  shapeProperties = NULL,
  placeholder = NULL
)
}
\arguments{
\item{shapeType}{(string)
The type of the shape.}

\item{text}{(\link{TextContent})
The text content of the shape.}

\item{shapeProperties}{(\link{ShapeProperties})
The properties of the shape.}

\item{placeholder}{(\link{Placeholder})
Placeholders are page elements that inherit from corresponding placeholders on layouts and masters. If set, the shape is a placeholder shape and any inherited properties can be resolved by looking at the parent placeholder identified by the Placeholder.parent_object_id field.}
}
\value{
Shape object
}
\description{
Shape Object
}
\details{
A PageElement kind representing a generic shape that does not have a more specific classification.
}
