% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ShapeProperties}
\alias{ShapeProperties}
\title{ShapeProperties Object}
\usage{
ShapeProperties(
  shapeBackgroundFill = NULL,
  autofit = NULL,
  outline = NULL,
  shadow = NULL,
  contentAlignment = NULL,
  link = NULL
)
}
\arguments{
\item{shapeBackgroundFill}{The background fill of the shape}

\item{autofit}{The autofit properties of the shape}

\item{outline}{The outline of the shape}

\item{shadow}{The shadow properties of the shape}

\item{contentAlignment}{The alignment of the content in the shape}

\item{link}{The hyperlink destination of the shape}
}
\value{
ShapeProperties object
}
\description{
ShapeProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of a Shape. If the shape is a placeholder shape as determined by the placeholder field, then these properties may be inherited from a parent placeholder shape. Determining the rendered value of the property depends on the corresponding property_state field value. Any text autofit settings on the shape are automatically deactivated by requests that can impact how text fits in the shape.
}
\concept{ShapeProperties functions}
