% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{SheetsChart}
\alias{SheetsChart}
\title{SheetsChart Object}
\usage{
SheetsChart(
  spreadsheetId = NULL,
  sheetsChartProperties = NULL,
  chartId = NULL,
  contentUrl = NULL
)
}
\arguments{
\item{spreadsheetId}{The ID of the Google Sheets spreadsheet that contains the source chart}

\item{sheetsChartProperties}{The properties of the Sheets chart}

\item{chartId}{The ID of the specific chart in the Google Sheets spreadsheet that is embedded}

\item{contentUrl}{The URL of an image of the embedded chart, with a default lifetime of 30 minutes}
}
\value{
SheetsChart object
}
\description{
SheetsChart Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A PageElement kind representing a linked chart embedded from Google Sheets.
}
\concept{SheetsChart functions}
