% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{SheetsChartProperties}
\alias{SheetsChartProperties}
\title{SheetsChartProperties Object}
\usage{
SheetsChartProperties(chartImageProperties = NULL)
}
\arguments{
\item{chartImageProperties}{(\link{ImageProperties})
The properties of the embedded chart image.}
}
\value{
SheetsChartProperties object
}
\description{
SheetsChartProperties Object
}
\details{
The properties of the SheetsChart.
}
