% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Size}
\alias{Size}
\title{Size Object}
\usage{
Size(width = NULL, height = NULL)
}
\arguments{
\item{width}{(\link{Dimension})
The width of the object.}

\item{height}{(\link{Dimension})
The height of the object.}
}
\value{
Size object
}
\description{
Size Object
}
\details{
A width and height.
}
