% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_auth.R
\name{slides_deauth}
\alias{slides_deauth}
\title{Suspend authorization}
\usage{
slides_deauth()
}
\description{
Put SlidesTools into a de-authorized state. Instead of sending a token,
SlidesTools will send an API key. This can be used to access public
resources for which no Google sign-in is required. This is handy for using
SlidesTools in a non-interactive setting to make requests that do not
require a token. It will prevent the attempt to obtain a token
interactively in the browser. The user can configure their own API key
via \code{\link[=slides_auth_configure]{slides_auth_configure()}} and retrieve that key via
\code{\link[=slides_api_key]{slides_api_key()}}.
In the absence of a user-configured key, a built-in default key is used.
}
\examples{
\dontrun{
slides_deauth()
slides_user()
}
}
\seealso{
Other auth functions: 
\code{\link{slides_auth}()},
\code{\link{slides_auth_configure}()}
}
\concept{auth functions}
