% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsd_auth.R
\name{slides_has_token}
\alias{slides_has_token}
\title{Is there a token on hand?}
\usage{
slides_has_token()
}
\value{
Logical.
}
\description{
Reports whether SlidesTools has stored a token, ready for use in downstream
requests.
}
\examples{
slides_has_token()
}
\seealso{
Other low-level API functions: 
\code{\link{slides_token}()}
}
\concept{low-level API functions}
