% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleslides.api-package.R
\docType{package}
\name{SlidesTools-package}
\alias{SlidesTools}
\alias{SlidesTools-package}
\title{SlidesTools: Google Slides API client and tools}
\description{
Create, read and modify Slides presentations with full REST API functionality.
}
\author{
\strong{Maintainer}: Isaac Gravestock \email{isaac.gravestock@gmail.com} [compiler]

Other contributors:
\itemize{
  \item Google (Createor of API and documentation) [copyright holder]
}

}
\keyword{internal}
