% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{SolidFill}
\alias{SolidFill}
\title{SolidFill Object}
\usage{
SolidFill(color = NULL, alpha = NULL)
}
\arguments{
\item{color}{The color value of the solid fill}

\item{alpha}{The fraction of this `color` that should be applied to the pixel}
}
\value{
SolidFill object
}
\description{
SolidFill Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A solid color fill. The page or page element is filled entirely with the specified color value. If any field is unset, its value may be inherited from a parent placeholder if it exists.
}
\concept{SolidFill functions}
