% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{SpeakerSpotlight}
\alias{SpeakerSpotlight}
\title{SpeakerSpotlight Object}
\usage{
SpeakerSpotlight(speakerSpotlightProperties = NULL)
}
\arguments{
\item{speakerSpotlightProperties}{(\link{SpeakerSpotlightProperties})
The properties of the Speaker Spotlight.}
}
\value{
SpeakerSpotlight object
}
\description{
SpeakerSpotlight Object
}
\details{
A PageElement kind representing a Speaker Spotlight.
}
