% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{SpeakerSpotlightProperties}
\alias{SpeakerSpotlightProperties}
\title{SpeakerSpotlightProperties Object}
\usage{
SpeakerSpotlightProperties(outline = NULL, shadow = NULL)
}
\arguments{
\item{outline}{(\link{Outline})
The outline of the Speaker Spotlight. If not set, it has no outline.}

\item{shadow}{(\link{Shadow})
The shadow of the Speaker Spotlight. If not set, it has no shadow.}
}
\value{
SpeakerSpotlightProperties object
}
\description{
SpeakerSpotlightProperties Object
}
\details{
The properties of the SpeakerSpotlight.
}
