% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{StretchedPictureFill}
\alias{StretchedPictureFill}
\title{StretchedPictureFill Object}
\usage{
StretchedPictureFill(size = NULL, contentUrl = NULL)
}
\arguments{
\item{size}{The original size of the picture fill}

\item{contentUrl}{Reading the content_url: An URL to a picture with a default lifetime of 30 minutes}
}
\value{
StretchedPictureFill object
}
\description{
StretchedPictureFill Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The stretched picture fill. The page or page element is filled entirely with the specified picture. The picture is stretched to fit its container.
}
\concept{StretchedPictureFill functions}
