% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{SubstringMatchCriteria}
\alias{SubstringMatchCriteria}
\title{SubstringMatchCriteria Object}
\usage{
SubstringMatchCriteria(text = NULL, matchCase = NULL)
}
\arguments{
\item{text}{(string)
The text to search for in the shape or table.}

\item{matchCase}{(boolean)
Indicates whether the search should respect case: - \code{True}: the search is case sensitive. - \code{False}: the search is case insensitive.}
}
\value{
SubstringMatchCriteria object
}
\description{
SubstringMatchCriteria Object
}
\details{
A criteria that matches a specific string of text in a shape or table.
}
