% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Table}
\alias{Table}
\title{Table Object}
\usage{
Table(
  columns = NULL,
  horizontalBorderRows = NULL,
  tableRows = NULL,
  verticalBorderRows = NULL,
  tableColumns = NULL,
  rows = NULL
)
}
\arguments{
\item{columns}{Number of columns in the table}

\item{horizontalBorderRows}{Properties of horizontal cell borders}

\item{tableRows}{Properties and contents of each row}

\item{verticalBorderRows}{Properties of vertical cell borders}

\item{tableColumns}{Properties of each column}

\item{rows}{Number of rows in the table}
}
\value{
Table object
}
\description{
Table Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A PageElement kind representing a table.
}
\concept{Table functions}
