% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableBorderCell}
\alias{TableBorderCell}
\title{TableBorderCell Object}
\usage{
TableBorderCell(location = NULL, tableBorderProperties = NULL)
}
\arguments{
\item{location}{(\link{TableCellLocation})
The location of the border within the border table.}

\item{tableBorderProperties}{(\link{TableBorderProperties})
The border properties.}
}
\value{
TableBorderCell object
}
\description{
TableBorderCell Object
}
\details{
The properties of each border cell.
}
