% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableBorderProperties}
\alias{TableBorderProperties}
\title{TableBorderProperties Object}
\usage{
TableBorderProperties(tableBorderFill = NULL, weight = NULL, dashStyle = NULL)
}
\arguments{
\item{tableBorderFill}{(\link{TableBorderFill})
The fill of the table border.}

\item{weight}{(\link{Dimension})
The thickness of the border.}

\item{dashStyle}{(string)
The dash style of the border.}
}
\value{
TableBorderProperties object
}
\description{
TableBorderProperties Object
}
\details{
The border styling properties of the TableBorderCell.
}
