% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableBorderRow}
\alias{TableBorderRow}
\title{TableBorderRow Object}
\usage{
TableBorderRow(tableBorderCells = NULL)
}
\arguments{
\item{tableBorderCells}{Properties of each border cell}
}
\value{
TableBorderRow object
}
\description{
TableBorderRow Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contents of each border row in a table.
}
\concept{TableBorderRow functions}
