% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableCell}
\alias{TableCell}
\title{TableCell Object}
\usage{
TableCell(
  location = NULL,
  tableCellProperties = NULL,
  rowSpan = NULL,
  text = NULL,
  columnSpan = NULL
)
}
\arguments{
\item{location}{The location of the cell within the table}

\item{tableCellProperties}{The properties of the table cell}

\item{rowSpan}{Row span of the cell}

\item{text}{The text content of the cell}

\item{columnSpan}{Column span of the cell}
}
\value{
TableCell object
}
\description{
TableCell Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties and contents of each table cell.
}
\concept{TableCell functions}
