% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableCellBackgroundFill}
\alias{TableCellBackgroundFill}
\title{TableCellBackgroundFill Object}
\usage{
TableCellBackgroundFill(propertyState = NULL, solidFill = NULL)
}
\arguments{
\item{propertyState}{The background fill property state}

\item{solidFill}{Solid color fill}
}
\value{
TableCellBackgroundFill object
}
\description{
TableCellBackgroundFill Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The table cell background fill.
}
\concept{TableCellBackgroundFill functions}
