% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableCellProperties}
\alias{TableCellProperties}
\title{TableCellProperties Object}
\usage{
TableCellProperties(tableCellBackgroundFill = NULL, contentAlignment = NULL)
}
\arguments{
\item{tableCellBackgroundFill}{(\link{TableCellBackgroundFill})
The background fill of the table cell. The default fill matches the fill for newly created table cells in the Slides editor.}

\item{contentAlignment}{(string)
The alignment of the content in the table cell. The default alignment matches the alignment for newly created table cells in the Slides editor.}
}
\value{
TableCellProperties object
}
\description{
TableCellProperties Object
}
\details{
The properties of the TableCell.
}
