% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableColumnProperties}
\alias{TableColumnProperties}
\title{TableColumnProperties Object}
\usage{
TableColumnProperties(columnWidth = NULL)
}
\arguments{
\item{columnWidth}{(\link{Dimension})
Width of a column.}
}
\value{
TableColumnProperties object
}
\description{
TableColumnProperties Object
}
\details{
Properties of each column in a table.
}
