% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableRange}
\alias{TableRange}
\title{TableRange Object}
\usage{
TableRange(location = NULL, rowSpan = NULL, columnSpan = NULL)
}
\arguments{
\item{location}{(\link{TableCellLocation})
The starting location of the table range.}

\item{rowSpan}{(integer)
The row span of the table range.}

\item{columnSpan}{(integer)
The column span of the table range.}
}
\value{
TableRange object
}
\description{
TableRange Object
}
\details{
A table range represents a reference to a subset of a table. It's important to note that the cells specified by a table range do not necessarily form a rectangle. For example, let's say we have a 3 x 3 table where all the cells of the last row are merged together. The table looks like this: [ ] A table range with location = (0, 0), row span = 3 and column span = 2 specifies the following cells: x x [ x x x ]

[ ]: R:\%20
[ x x x ]: R:\%20x\%20x\%20x\%20
}
