% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableRowProperties}
\alias{TableRowProperties}
\title{TableRowProperties Object}
\usage{
TableRowProperties(minRowHeight = NULL)
}
\arguments{
\item{minRowHeight}{Minimum height of the row}
}
\value{
TableRowProperties object
}
\description{
TableRowProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties of each row in a table.
}
\concept{TableRowProperties functions}
