% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TextContent}
\alias{TextContent}
\title{TextContent Object}
\usage{
TextContent(textElements = NULL, lists = NULL)
}
\arguments{
\item{textElements}{(list of \link{TextElement} objects)
The text contents broken down into its component parts, including styling information. This property is read-only.}

\item{lists}{(named list of \link{List} objects)
The bulleted lists contained in this text, keyed by list ID.}
}
\value{
TextContent object
}
\description{
TextContent Object
}
\details{
The general text content. The text must reside in a compatible shape (e.g. text box or rectangle) or a table cell in a page.
}
