% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TextElement}
\alias{TextElement}
\title{TextElement Object}
\usage{
TextElement(
  startIndex = NULL,
  endIndex = NULL,
  paragraphMarker = NULL,
  textRun = NULL,
  autoText = NULL
)
}
\arguments{
\item{startIndex}{(integer)
The zero-based start index of this text element, in Unicode code units.}

\item{endIndex}{(integer)
The zero-based end index of this text element, exclusive, in Unicode code units.}

\item{paragraphMarker}{(\link{ParagraphMarker})
A marker representing the beginning of a new paragraph. The \code{start_index} and \code{end_index} of this TextElement represent the range of the paragraph. Other TextElements with an index range contained inside this paragraph's range are considered to be part of this paragraph. The range of indices of two separate paragraphs will never overlap.}

\item{textRun}{(\link{TextRun})
A TextElement representing a run of text where all of the characters in the run have the same TextStyle. The \code{start_index} and \code{end_index} of TextRuns will always be fully contained in the index range of a single \code{paragraph_marker} TextElement. In other words, a TextRun will never span multiple paragraphs.}

\item{autoText}{(\link{AutoText})
A TextElement representing a spot in the text that is dynamically replaced with content that can change over time.}
}
\value{
TextElement object
}
\description{
TextElement Object
}
\details{
A TextElement describes the content of a range of indices in the text content of a Shape or TableCell.
}
