% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TextRun}
\alias{TextRun}
\title{TextRun Object}
\usage{
TextRun(content = NULL, style = NULL)
}
\arguments{
\item{content}{(string)
The text of this run.}

\item{style}{(\link{TextStyle})
The styling applied to this run.}
}
\value{
TextRun object
}
\description{
TextRun Object
}
\details{
A TextElement kind that represents a run of text that all has the same styling.
}
