% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ThemeColorPair}
\alias{ThemeColorPair}
\title{ThemeColorPair Object}
\usage{
ThemeColorPair(type = NULL, color = NULL)
}
\arguments{
\item{type}{(string)
The type of the theme color.
\itemize{
\item \code{THEME_COLOR_TYPE_UNSPECIFIED}: Unspecified theme color. This value should not be used.
\item \code{DARK1}: Represents the first dark color.
\item \code{LIGHT1}: Represents the first light color.
\item \code{DARK2}: Represents the second dark color.
\item \code{LIGHT2}: Represents the second light color.
\item \code{ACCENT1}: Represents the first accent color.
\item \code{ACCENT2}: Represents the second accent color.
\item \code{ACCENT3}: Represents the third accent color.
\item \code{ACCENT4}: Represents the fourth accent color.
\item \code{ACCENT5}: Represents the fifth accent color.
\item \code{ACCENT6}: Represents the sixth accent color.
\item \code{HYPERLINK}: Represents the color to use for hyperlinks.
\item \code{FOLLOWED_HYPERLINK}: Represents the color to use for visited hyperlinks.
\item \code{TEXT1}: Represents the first text color.
\item \code{BACKGROUND1}: Represents the first background color.
\item \code{TEXT2}: Represents the second text color.
\item \code{BACKGROUND2}: Represents the second background color.
}}

\item{color}{(\link{RgbColor})
The concrete color corresponding to the theme color type above.}
}
\value{
ThemeColorPair object
}
\description{
ThemeColorPair Object
}
\details{
A pair mapping a theme color type to the concrete color it represents.
}
