% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Thumbnail}
\alias{Thumbnail}
\title{Thumbnail Object}
\usage{
Thumbnail(height = NULL, contentUrl = NULL, width = NULL)
}
\arguments{
\item{height}{The positive height in pixels of the thumbnail image}

\item{contentUrl}{The content URL of the thumbnail image}

\item{width}{The positive width in pixels of the thumbnail image}
}
\value{
Thumbnail object
}
\description{
Thumbnail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The thumbnail of a page.
}
\concept{Thumbnail functions}
