% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UngroupObjectsRequest}
\alias{UngroupObjectsRequest}
\title{UngroupObjectsRequest Object}
\usage{
UngroupObjectsRequest(objectIds = NULL)
}
\arguments{
\item{objectIds}{The object IDs of the objects to ungroup}
}
\value{
UngroupObjectsRequest object
}
\description{
UngroupObjectsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Ungroups objects, such as groups.
}
\concept{UngroupObjectsRequest functions}
