% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UnmergeTableCellsRequest}
\alias{UnmergeTableCellsRequest}
\title{UnmergeTableCellsRequest Object}
\usage{
UnmergeTableCellsRequest(objectId = NULL, tableRange = NULL)
}
\arguments{
\item{objectId}{(string)
The object ID of the table.}

\item{tableRange}{(\link{TableRange})
The table range specifying which cells of the table to unmerge. All merged cells in this range will be unmerged, and cells that are already unmerged will not be affected. If the range has no merged cells, the request will do nothing. If there is text in any of the merged cells, the text will remain in the upper-left ("head") cell of the resulting block of unmerged cells.}
}
\value{
UnmergeTableCellsRequest object
}
\description{
UnmergeTableCellsRequest Object
}
\details{
Unmerges cells in a Table.
}
