% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateImagePropertiesRequest}
\alias{UpdateImagePropertiesRequest}
\title{UpdateImagePropertiesRequest Object}
\usage{
UpdateImagePropertiesRequest(
  objectId = NULL,
  imageProperties = NULL,
  fields = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID of the image the updates are applied to.}

\item{imageProperties}{(\link{ImageProperties})
The image properties to update.}

\item{fields}{(string)
The fields that should be updated. At least one field must be specified. The root \code{imageProperties} is implied and should not be specified. A single \code{"*"} can be used as short-hand for listing every field. For example to update the image outline color, set \code{fields} to \code{"outline.outlineFill.solidFill.color"}. To reset a property to its default value, include its field name in the field mask but leave the field itself unset.}
}
\value{
UpdateImagePropertiesRequest object
}
\description{
UpdateImagePropertiesRequest Object
}
\details{
Update the properties of an Image.
}
