% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateLineCategoryRequest}
\alias{UpdateLineCategoryRequest}
\title{UpdateLineCategoryRequest Object}
\usage{
UpdateLineCategoryRequest(objectId = NULL, lineCategory = NULL)
}
\arguments{
\item{objectId}{(string)
The object ID of the line the update is applied to. Only a line with a category indicating it is a "connector" can be updated. The line may be rerouted after updating its category.}

\item{lineCategory}{(string)
The line category to update to. The exact line type is determined based on the category to update to and how it's routed to connect to other page elements.}
}
\value{
UpdateLineCategoryRequest object
}
\description{
UpdateLineCategoryRequest Object
}
\details{
Updates the category of a line.
}
