% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateLinePropertiesRequest}
\alias{UpdateLinePropertiesRequest}
\title{UpdateLinePropertiesRequest Object}
\usage{
UpdateLinePropertiesRequest(
  fields = NULL,
  objectId = NULL,
  lineProperties = NULL
)
}
\arguments{
\item{fields}{The fields that should be updated}

\item{objectId}{The object ID of the line the update is applied to}

\item{lineProperties}{The line properties to update}
}
\value{
UpdateLinePropertiesRequest object
}
\description{
UpdateLinePropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the properties of a Line.
}
\concept{UpdateLinePropertiesRequest functions}
