% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdatePageElementAltTextRequest}
\alias{UpdatePageElementAltTextRequest}
\title{UpdatePageElementAltTextRequest Object}
\usage{
UpdatePageElementAltTextRequest(
  objectId = NULL,
  title = NULL,
  description = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID of the page element the updates are applied to.}

\item{title}{(string)
The updated alt text title of the page element. If unset the existing value will be maintained. The title is exposed to screen readers and other accessibility interfaces. Only use human readable values related to the content of the page element.}

\item{description}{(string)
The updated alt text description of the page element. If unset the existing value will be maintained. The description is exposed to screen readers and other accessibility interfaces. Only use human readable values related to the content of the page element.}
}
\value{
UpdatePageElementAltTextRequest object
}
\description{
UpdatePageElementAltTextRequest Object
}
\details{
Updates the alt text title and/or description of a page element.
}
