% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdatePageElementsZOrderRequest}
\alias{UpdatePageElementsZOrderRequest}
\title{UpdatePageElementsZOrderRequest Object}
\usage{
UpdatePageElementsZOrderRequest(operation = NULL, pageElementObjectIds = NULL)
}
\arguments{
\item{operation}{The Z-order operation to apply on the page elements}

\item{pageElementObjectIds}{The object IDs of the page elements to update}
}
\value{
UpdatePageElementsZOrderRequest object
}
\description{
UpdatePageElementsZOrderRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the Z-order of page elements. Z-order is an ordering of the elements on the page from back to front. The page element in the front may cover the elements that are behind it.
}
\concept{UpdatePageElementsZOrderRequest functions}
