% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdatePageElementTransformRequest}
\alias{UpdatePageElementTransformRequest}
\title{UpdatePageElementTransformRequest Object}
\usage{
UpdatePageElementTransformRequest(
  transform = NULL,
  objectId = NULL,
  applyMode = NULL
)
}
\arguments{
\item{transform}{The input transform matrix used to update the page element}

\item{objectId}{The object ID of the page element to update}

\item{applyMode}{The apply mode of the transform update}
}
\value{
UpdatePageElementTransformRequest object
}
\description{
UpdatePageElementTransformRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the transform of a page element. Updating the transform of a group will change the absolute transform of the page elements in that group, which can change their visual appearance. See the documentation for PageElement.transform for more details.
}
\concept{UpdatePageElementTransformRequest functions}
