% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdatePagePropertiesRequest}
\alias{UpdatePagePropertiesRequest}
\title{UpdatePagePropertiesRequest Object}
\usage{
UpdatePagePropertiesRequest(
  objectId = NULL,
  pageProperties = NULL,
  fields = NULL
)
}
\arguments{
\item{objectId}{The object ID of the page the update is applied to}

\item{pageProperties}{The page properties to update}

\item{fields}{The fields that should be updated}
}
\value{
UpdatePagePropertiesRequest object
}
\description{
UpdatePagePropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the properties of a Page.
}
\concept{UpdatePagePropertiesRequest functions}
