% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateShapePropertiesRequest}
\alias{UpdateShapePropertiesRequest}
\title{UpdateShapePropertiesRequest Object}
\usage{
UpdateShapePropertiesRequest(
  objectId = NULL,
  shapeProperties = NULL,
  fields = NULL
)
}
\arguments{
\item{objectId}{The object ID of the shape the updates are applied to}

\item{shapeProperties}{The shape properties to update}

\item{fields}{The fields that should be updated}
}
\value{
UpdateShapePropertiesRequest object
}
\description{
UpdateShapePropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Update the properties of a Shape.
}
\concept{UpdateShapePropertiesRequest functions}
