% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateSlidePropertiesRequest}
\alias{UpdateSlidePropertiesRequest}
\title{UpdateSlidePropertiesRequest Object}
\usage{
UpdateSlidePropertiesRequest(
  objectId = NULL,
  slideProperties = NULL,
  fields = NULL
)
}
\arguments{
\item{objectId}{The object ID of the slide the update is applied to}

\item{slideProperties}{The slide properties to update}

\item{fields}{The fields that should be updated}
}
\value{
UpdateSlidePropertiesRequest object
}
\description{
UpdateSlidePropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the properties of a Slide.
}
\concept{UpdateSlidePropertiesRequest functions}
