% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateSlidesPositionRequest}
\alias{UpdateSlidesPositionRequest}
\title{UpdateSlidesPositionRequest Object}
\usage{
UpdateSlidesPositionRequest(slideObjectIds = NULL, insertionIndex = NULL)
}
\arguments{
\item{slideObjectIds}{(list of \link{string} objects)
The IDs of the slides in the presentation that should be moved. The slides in this list must be in existing presentation order, without duplicates.}

\item{insertionIndex}{(integer)
The index where the slides should be inserted, based on the slide arrangement before the move takes place. Must be between zero and the number of slides in the presentation, inclusive.}
}
\value{
UpdateSlidesPositionRequest object
}
\description{
UpdateSlidesPositionRequest Object
}
\details{
Updates the position of slides in the presentation.
}
