% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateTableBorderPropertiesRequest}
\alias{UpdateTableBorderPropertiesRequest}
\title{UpdateTableBorderPropertiesRequest Object}
\usage{
UpdateTableBorderPropertiesRequest(
  objectId = NULL,
  tableRange = NULL,
  borderPosition = NULL,
  tableBorderProperties = NULL,
  fields = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID of the table.}

\item{tableRange}{(\link{TableRange})
The table range representing the subset of the table to which the updates are applied. If a table range is not specified, the updates will apply to the entire table.}

\item{borderPosition}{(string)
The border position in the table range the updates should apply to. If a border position is not specified, the updates will apply to all borders in the table range.
\itemize{
\item \code{ALL}: All borders in the range.
\item \code{BOTTOM}: Borders at the bottom of the range.
\item \code{INNER}: Borders on the inside of the range.
\item \code{INNER_HORIZONTAL}: Horizontal borders on the inside of the range.
\item \code{INNER_VERTICAL}: Vertical borders on the inside of the range.
\item \code{LEFT}: Borders at the left of the range.
\item \code{OUTER}: Borders along the outside of the range.
\item \code{RIGHT}: Borders at the right of the range.
\item \code{TOP}: Borders at the top of the range.
}}

\item{tableBorderProperties}{(\link{TableBorderProperties})
The table border properties to update.}

\item{fields}{(string)
The fields that should be updated. At least one field must be specified. The root \code{tableBorderProperties} is implied and should not be specified. A single \code{"*"} can be used as short-hand for listing every field. For example to update the table border solid fill color, set \code{fields} to \code{"tableBorderFill.solidFill.color"}. To reset a property to its default value, include its field name in the field mask but leave the field itself unset.}
}
\value{
UpdateTableBorderPropertiesRequest object
}
\description{
UpdateTableBorderPropertiesRequest Object
}
\details{
Updates the properties of the table borders in a Table.
}
