% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateTableCellPropertiesRequest}
\alias{UpdateTableCellPropertiesRequest}
\title{UpdateTableCellPropertiesRequest Object}
\usage{
UpdateTableCellPropertiesRequest(
  objectId = NULL,
  tableRange = NULL,
  tableCellProperties = NULL,
  fields = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID of the table.}

\item{tableRange}{(\link{TableRange})
The table range representing the subset of the table to which the updates are applied. If a table range is not specified, the updates will apply to the entire table.}

\item{tableCellProperties}{(\link{TableCellProperties})
The table cell properties to update.}

\item{fields}{(string)
The fields that should be updated. At least one field must be specified. The root \code{tableCellProperties} is implied and should not be specified. A single \code{"*"} can be used as short-hand for listing every field. For example to update the table cell background solid fill color, set \code{fields} to \code{"tableCellBackgroundFill.solidFill.color"}. To reset a property to its default value, include its field name in the field mask but leave the field itself unset.}
}
\value{
UpdateTableCellPropertiesRequest object
}
\description{
UpdateTableCellPropertiesRequest Object
}
\details{
Update the properties of a TableCell.
}
