% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateTableColumnPropertiesRequest}
\alias{UpdateTableColumnPropertiesRequest}
\title{UpdateTableColumnPropertiesRequest Object}
\usage{
UpdateTableColumnPropertiesRequest(
  fields = NULL,
  tableColumnProperties = NULL,
  columnIndices = NULL,
  objectId = NULL
)
}
\arguments{
\item{fields}{The fields that should be updated}

\item{tableColumnProperties}{The table column properties to update}

\item{columnIndices}{The list of zero-based indices specifying which columns to update}

\item{objectId}{The object ID of the table}
}
\value{
UpdateTableColumnPropertiesRequest object
}
\description{
UpdateTableColumnPropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the properties of a Table column.
}
\concept{UpdateTableColumnPropertiesRequest functions}
