% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateTableRowPropertiesRequest}
\alias{UpdateTableRowPropertiesRequest}
\title{UpdateTableRowPropertiesRequest Object}
\usage{
UpdateTableRowPropertiesRequest(
  objectId = NULL,
  rowIndices = NULL,
  tableRowProperties = NULL,
  fields = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID of the table.}

\item{rowIndices}{(list of \link{integer} objects)
The list of zero-based indices specifying which rows to update. If no indices are provided, all rows in the table will be updated.}

\item{tableRowProperties}{(\link{TableRowProperties})
The table row properties to update.}

\item{fields}{(string)
The fields that should be updated. At least one field must be specified. The root \code{tableRowProperties} is implied and should not be specified. A single \code{"*"} can be used as short-hand for listing every field. For example to update the minimum row height, set \code{fields} to \code{"min_row_height"}. If '"min_row_height"' is included in the field mask but the property is left unset, the minimum row height will default to 0.}
}
\value{
UpdateTableRowPropertiesRequest object
}
\description{
UpdateTableRowPropertiesRequest Object
}
\details{
Updates the properties of a Table row.
}
