% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateTextStyleRequest}
\alias{UpdateTextStyleRequest}
\title{UpdateTextStyleRequest Object}
\usage{
UpdateTextStyleRequest(
  objectId = NULL,
  cellLocation = NULL,
  style = NULL,
  textRange = NULL,
  fields = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID of the shape or table with the text to be styled.}

\item{cellLocation}{(\link{TableCellLocation})
The location of the cell in the table containing the text to style. If \code{object_id} refers to a table, \code{cell_location} must have a value. Otherwise, it must not.}

\item{style}{(\link{TextStyle})
The style(s) to set on the text. If the value for a particular style matches that of the parent, that style will be set to inherit. Certain text style changes may cause other changes meant to mirror the behavior of the Slides editor. See the documentation of TextStyle for more information.}

\item{textRange}{(\link{Range})
The range of text to style. The range may be extended to include adjacent newlines. If the range fully contains a paragraph belonging to a list, the paragraph's bullet is also updated with the matching text style.}

\item{fields}{(string)
The fields that should be updated. At least one field must be specified. The root \code{style} is implied and should not be specified. A single \code{"*"} can be used as short-hand for listing every field. For example, to update the text style to bold, set \code{fields} to \code{"bold"}. To reset a property to its default value, include its field name in the field mask but leave the field itself unset.}
}
\value{
UpdateTextStyleRequest object
}
\description{
UpdateTextStyleRequest Object
}
\details{
Update the styling of text in a Shape or Table.
}
