% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateVideoPropertiesRequest}
\alias{UpdateVideoPropertiesRequest}
\title{UpdateVideoPropertiesRequest Object}
\usage{
UpdateVideoPropertiesRequest(
  objectId = NULL,
  videoProperties = NULL,
  fields = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID of the video the updates are applied to.}

\item{videoProperties}{(\link{VideoProperties})
The video properties to update.}

\item{fields}{(string)
The fields that should be updated. At least one field must be specified. The root \code{videoProperties} is implied and should not be specified. A single \code{"*"} can be used as short-hand for listing every field. For example to update the video outline color, set \code{fields} to \code{"outline.outlineFill.solidFill.color"}. To reset a property to its default value, include its field name in the field mask but leave the field itself unset.}
}
\value{
UpdateVideoPropertiesRequest object
}
\description{
UpdateVideoPropertiesRequest Object
}
\details{
Update the properties of a Video.
}
