% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{WordArt}
\alias{WordArt}
\title{WordArt Object}
\usage{
WordArt(renderedText = NULL)
}
\arguments{
\item{renderedText}{The text rendered as word art}
}
\value{
WordArt object
}
\description{
WordArt Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A PageElement kind representing word art.
}
\concept{WordArt functions}
