% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{WriteControl}
\alias{WriteControl}
\title{WriteControl Object}
\usage{
WriteControl(requiredRevisionId = NULL)
}
\arguments{
\item{requiredRevisionId}{The revision ID of the presentation required for the write request}
}
\value{
WriteControl object
}
\description{
WriteControl Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Provides control over how write requests are executed.
}
\concept{WriteControl functions}
