% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{AffineTransform}
\alias{AffineTransform}
\title{AffineTransform Object}
\usage{
AffineTransform(
  scaleX = NULL,
  scaleY = NULL,
  shearX = NULL,
  shearY = NULL,
  translateX = NULL,
  translateY = NULL,
  unit = NULL
)
}
\arguments{
\item{scaleX}{(number)
The X coordinate scaling element.}

\item{scaleY}{(number)
The Y coordinate scaling element.}

\item{shearX}{(number)
The X coordinate shearing element.}

\item{shearY}{(number)
The Y coordinate shearing element.}

\item{translateX}{(number)
The X coordinate translation element.}

\item{translateY}{(number)
The Y coordinate translation element.}

\item{unit}{(string)
The units for translate elements.
\itemize{
\item \code{UNIT_UNSPECIFIED}: The units are unknown.
\item \code{EMU}: An English Metric Unit (EMU) is defined as 1/360,000 of a centimeter and thus there are 914,400 EMUs per inch, and 12,700 EMUs per point.
\item \code{PT}: A point, 1/72 of an inch.
}}
}
\value{
AffineTransform object
}
\description{
AffineTransform Object
}
\details{
AffineTransform uses a 3x3 matrix with an implied last row of [ 0 0 1 ] to transform source coordinates (x,y) into destination coordinates (x', y') according to: x' x = shear_y scale_y translate_y 1 [ 1 ] After transformation, x' = scale_x * x + shear_x * y + translate_x; y' = scale_y * y + shear_y * x + translate_y; This message is therefore composed of these six matrix elements.
}
