% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{BatchUpdatePresentationRequest}
\alias{BatchUpdatePresentationRequest}
\title{BatchUpdatePresentationRequest Object}
\usage{
BatchUpdatePresentationRequest(writeControl = NULL, requests = NULL)
}
\arguments{
\item{writeControl}{Provides control over how write requests are executed}

\item{requests}{A list of updates to apply to the presentation}
}
\value{
BatchUpdatePresentationRequest object
}
\description{
BatchUpdatePresentationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for PresentationsService.BatchUpdatePresentation.
}
\seealso{
Other BatchUpdatePresentationRequest functions: 
\code{\link{presentations.batchUpdate}()}
}
\concept{BatchUpdatePresentationRequest functions}
