% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{BatchUpdatePresentationResponse}
\alias{BatchUpdatePresentationResponse}
\title{BatchUpdatePresentationResponse Object}
\usage{
BatchUpdatePresentationResponse(
  replies = NULL,
  writeControl = NULL,
  presentationId = NULL
)
}
\arguments{
\item{replies}{The reply of the updates}

\item{writeControl}{The updated write control after applying the request}

\item{presentationId}{The presentation the updates were applied to}
}
\value{
BatchUpdatePresentationResponse object
}
\description{
BatchUpdatePresentationResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message from a batch update.
}
\concept{BatchUpdatePresentationResponse functions}
