% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ColorStop}
\alias{ColorStop}
\title{ColorStop Object}
\usage{
ColorStop(color = NULL, alpha = NULL, position = NULL)
}
\arguments{
\item{color}{(\link{OpaqueColor})
The color of the gradient stop.}

\item{alpha}{(number)
The alpha value of this color in the gradient band. Defaults to 1.0, fully opaque.}

\item{position}{(number)
The relative position of the color stop in the gradient band measured in percentage. The value should be in the interval [0.0, 1.0].}
}
\value{
ColorStop object
}
\description{
ColorStop Object
}
\details{
A color and position in a gradient band.
}
