% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateImageRequest}
\alias{CreateImageRequest}
\title{CreateImageRequest Object}
\usage{
CreateImageRequest(elementProperties = NULL, objectId = NULL, url = NULL)
}
\arguments{
\item{elementProperties}{The element properties for the image}

\item{objectId}{A user-supplied object ID}

\item{url}{The image URL}
}
\value{
CreateImageRequest object
}
\description{
CreateImageRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creates an image.
}
\concept{CreateImageRequest functions}
