% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateShapeResponse}
\alias{CreateShapeResponse}
\title{CreateShapeResponse Object}
\usage{
CreateShapeResponse(objectId = NULL)
}
\arguments{
\item{objectId}{(string)
The object ID of the created shape.}
}
\value{
CreateShapeResponse object
}
\description{
CreateShapeResponse Object
}
\details{
The result of creating a shape.
}
