% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateSheetsChartRequest}
\alias{CreateSheetsChartRequest}
\title{CreateSheetsChartRequest Object}
\usage{
CreateSheetsChartRequest(
  objectId = NULL,
  elementProperties = NULL,
  spreadsheetId = NULL,
  chartId = NULL,
  linkingMode = NULL
)
}
\arguments{
\item{objectId}{(string)
A user-supplied object ID. If specified, the ID must be unique among all pages and page elements in the presentation. The ID should start with a word character [a-zA-Z0-9_] and then followed by any number of the following characters [a-zA-Z0-9_-:]. The length of the ID should not be less than 5 or greater than 50. If empty, a unique identifier will be generated.}

\item{elementProperties}{(\link{PageElementProperties})
The element properties for the chart. When the aspect ratio of the provided size does not match the chart aspect ratio, the chart is scaled and centered with respect to the size in order to maintain aspect ratio. The provided transform is applied after this operation.}

\item{spreadsheetId}{(string)
The ID of the Google Sheets spreadsheet that contains the chart. You might need to add a resource key to the HTTP header for a subset of old files. For more information, see \href{https://developers.google.com/drive/api/v3/resource-keys}{Access link-shared files using resource keys}.}

\item{chartId}{(integer)
The ID of the specific chart in the Google Sheets spreadsheet.}

\item{linkingMode}{(string)
The mode with which the chart is linked to the source spreadsheet. When not specified, the chart will be an image that is not linked.}
}
\value{
CreateSheetsChartRequest object
}
\description{
CreateSheetsChartRequest Object
}
\details{
Creates an embedded Google Sheets chart. NOTE: Chart creation requires at least one of the spreadsheets.readonly, spreadsheets, drive.readonly, drive.file, or drive OAuth scopes.
}
