% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{CreateVideoRequest}
\alias{CreateVideoRequest}
\title{CreateVideoRequest Object}
\usage{
CreateVideoRequest(
  objectId = NULL,
  elementProperties = NULL,
  source = NULL,
  id = NULL
)
}
\arguments{
\item{objectId}{(string)
A user-supplied object ID. If you specify an ID, it must be unique among all pages and page elements in the presentation. The ID must start with an alphanumeric character or an underscore (matches regex \verb{[a-zA-Z0-9_]}); remaining characters may include those as well as a hyphen or colon (matches regex \verb{[a-zA-Z0-9_-:]}). The length of the ID must not be less than 5 or greater than 50. If you don't specify an ID, a unique one is generated.}

\item{elementProperties}{(\link{PageElementProperties})
The element properties for the video. The PageElementProperties.size property is optional. If you don't specify a size, a default size is chosen by the server. The PageElementProperties.transform property is optional. The transform must not have shear components. If you don't specify a transform, the video will be placed at the top left corner of the page.}

\item{source}{(string)
The video source.
\itemize{
\item \code{SOURCE_UNSPECIFIED}: The video source is unspecified.
\item \code{YOUTUBE}: The video source is YouTube.
\item \code{DRIVE}: The video source is Google Drive.
}}

\item{id}{(string)
The video source's unique identifier for this video. e.g. For YouTube video https://www.youtube.com/watch?v=7U3axjORYZ0, the ID is 7U3axjORYZ0. For a Google Drive video https://drive.google.com/file/d/1xCgQLFTJi5_Xl8DgW_lcUYq5e-q6Hi5Q the ID is 1xCgQLFTJi5_Xl8DgW_lcUYq5e-q6Hi5Q. To access a Google Drive video file, you might need to add a resource key to the HTTP header for a subset of old files. For more information, see \href{https://developers.google.com/drive/api/v3/resource-keys}{Access link-shared files using resource keys}.}
}
\value{
CreateVideoRequest object
}
\description{
CreateVideoRequest Object
}
\details{
Creates a video. NOTE: Creating a video from Google Drive requires that the requesting app have at least one of the drive, drive.readonly, or drive.file OAuth scopes.
}
