% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{DeleteTableColumnRequest}
\alias{DeleteTableColumnRequest}
\title{DeleteTableColumnRequest Object}
\usage{
DeleteTableColumnRequest(tableObjectId = NULL, cellLocation = NULL)
}
\arguments{
\item{tableObjectId}{The table to delete columns from}

\item{cellLocation}{The reference table cell location from which a column will be deleted}
}
\value{
DeleteTableColumnRequest object
}
\description{
DeleteTableColumnRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deletes a column from a table.
}
\concept{DeleteTableColumnRequest functions}
