% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{GroupObjectsResponse}
\alias{GroupObjectsResponse}
\title{GroupObjectsResponse Object}
\usage{
GroupObjectsResponse(objectId = NULL)
}
\arguments{
\item{objectId}{(string)
The object ID of the created group.}
}
\value{
GroupObjectsResponse object
}
\description{
GroupObjectsResponse Object
}
\details{
The result of grouping objects.
}
