% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsd_auth.R
\name{gsd_auth_configure}
\alias{gsd_auth_configure}
\alias{gsd_oauth_app}
\title{Edit and view auth configuration}
\usage{
gsd_auth_configure(app, path)

gsd_oauth_app()
}
\arguments{
\item{app}{`r lifecycle::badge('deprecated')` Replaced by the `client`
argument.}

\item{path}{JSON downloaded from \href{https://console.cloud.google.com}{Google Cloud Console}, containing a client id and
secret, in one of the forms supported for the \code{txt} argument of
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}} (typically, a file path or JSON string).}

\item{client}{A Google OAuth client, presumably constructed via
[gargle::gargle_oauth_client_from_json()]. Note, however, that it is
preferred to specify the client with JSON, using the `path` argument.}
}
\value{
* `gsd_auth_configure()`: An object of R6 class
    [gargle::AuthState], invisibly.
  * `gsd_oauth_client()`: the current user-configured OAuth client.
}
\description{
These functions give more control over and visibility into the auth
configuration than [gsd_auth()] does. `gsd_auth_configure()`
lets the user specify their own:
  * OAuth client, which is used when obtaining a user token.

See the `vignette("get-api-credentials", package = "gargle")`
for more.
If the user does not configure these settings, internal defaults
are used.

`gsd_oauth_client()` retrieves the currently configured OAuth client.
}
\examples{
# see the current user-configured OAuth app (probaby `NULL`)
gsd_oauth_app()

if (require(httr)) {
  # store current state, so we can restore
  original_app <- gsd_oauth_app()

  # bring your own app via client id (aka key) and secret
  google_app <- httr::oauth_app(
    "my-awesome-google-api-wrapping-package",
    key = "123456789.apps.googleusercontent.com",
    secret = "abcdefghijklmnopqrstuvwxyz"
  )
  gsd_auth_configure(app = google_app)

  # confirm current app
  gsd_oauth_app()

  # restore original state
  gsd_auth_configure(app = original_app)
  gsd_oauth_app()
}

\dontrun{
# bring your own app via JSON downloaded from GCP Console
gsd_auth_configure(
  path = "/path/to/the/JSON/you/downloaded/from/gcp/console.json"
)
}

}
\seealso{
Other auth functions: 
\code{\link{gsd_auth}()},
\code{\link{gsd_deauth}()}
}
\concept{auth functions}
