% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Image}
\alias{Image}
\title{Image Object}
\usage{
Image(
  contentUrl = NULL,
  sourceUrl = NULL,
  imageProperties = NULL,
  placeholder = NULL
)
}
\arguments{
\item{contentUrl}{An URL to an image with a default lifetime of 30 minutes}

\item{sourceUrl}{The source URL is the URL used to insert the image}

\item{imageProperties}{The properties of the image}

\item{placeholder}{Placeholders are page elements that inherit from corresponding placeholders on layouts and masters}
}
\value{
Image object
}
\description{
Image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A PageElement kind representing an image.
}
\concept{Image functions}
