% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{LayoutReference}
\alias{LayoutReference}
\title{LayoutReference Object}
\usage{
LayoutReference(predefinedLayout = NULL, layoutId = NULL)
}
\arguments{
\item{predefinedLayout}{(string)
Predefined layout.}

\item{layoutId}{(string)
Layout ID: the object ID of one of the layouts in the presentation.}
}
\value{
LayoutReference object
}
\description{
LayoutReference Object
}
\details{
Slide layout reference. This may reference either: - A predefined layout - One of the layouts in the presentation.
}
