% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{LayoutReference}
\alias{LayoutReference}
\title{LayoutReference Object}
\usage{
LayoutReference(layoutId = NULL, predefinedLayout = NULL)
}
\arguments{
\item{layoutId}{Layout ID: the object ID of one of the layouts in the presentation}

\item{predefinedLayout}{Predefined layout}
}
\value{
LayoutReference object
}
\description{
LayoutReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Slide layout reference. This may reference either: - A predefined layout - One of the layouts in the presentation.
}
\concept{LayoutReference functions}
