% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Line}
\alias{Line}
\title{Line Object}
\usage{
Line(lineProperties = NULL, lineType = NULL, lineCategory = NULL)
}
\arguments{
\item{lineProperties}{(\link{LineProperties})
The properties of the line.}

\item{lineType}{(string)
The type of the line.}

\item{lineCategory}{(string)
The category of the line. It matches the \code{category} specified in CreateLineRequest, and can be updated with UpdateLineCategoryRequest.}
}
\value{
Line object
}
\description{
Line Object
}
\details{
A PageElement kind representing a non-connector line, straight connector, curved connector, or bent connector.
}
