% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{LineConnection}
\alias{LineConnection}
\title{LineConnection Object}
\usage{
LineConnection(connectedObjectId = NULL, connectionSiteIndex = NULL)
}
\arguments{
\item{connectedObjectId}{(string)
The object ID of the connected page element. Some page elements, such as groups, tables, and lines do not have connection sites and therefore cannot be connected to a connector line.}

\item{connectionSiteIndex}{(integer)
The index of the connection site on the connected page element. In most cases, it corresponds to the predefined connection site index from the ECMA-376 standard. More information on those connection sites can be found in the description of the "cnx" attribute in section 20.1.9.9 and Annex H. "Predefined DrawingML Shape and Text Geometries" of "Office Open XML File Formats-Fundamentals and Markup Language Reference", part 1 of \link{ECMA-376 5th edition} (http://www.ecma-international.org/publications/standards/Ecma-376.htm). The position of each connection site can also be viewed from Slides editor.}
}
\value{
LineConnection object
}
\description{
LineConnection Object
}
\details{
The properties for one end of a Line connection.
}
