% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{LineProperties}
\alias{LineProperties}
\title{LineProperties Object}
\usage{
LineProperties(
  lineFill = NULL,
  weight = NULL,
  dashStyle = NULL,
  startArrow = NULL,
  endArrow = NULL,
  link = NULL,
  startConnection = NULL,
  endConnection = NULL
)
}
\arguments{
\item{lineFill}{(\link{LineFill})
The fill of the line. The default line fill matches the defaults for new lines created in the Slides editor.}

\item{weight}{(\link{Dimension})
The thickness of the line.}

\item{dashStyle}{(string)
The dash style of the line.}

\item{startArrow}{(string)
The style of the arrow at the beginning of the line.}

\item{endArrow}{(string)
The style of the arrow at the end of the line.}

\item{link}{(\link{Link})
The hyperlink destination of the line. If unset, there is no link.}

\item{startConnection}{(\link{LineConnection})
The connection at the beginning of the line. If unset, there is no connection. Only lines with a Type indicating it is a "connector" can have a \code{start_connection}.}

\item{endConnection}{(\link{LineConnection})
The connection at the end of the line. If unset, there is no connection. Only lines with a Type indicating it is a "connector" can have an \code{end_connection}.}
}
\value{
LineProperties object
}
\description{
LineProperties Object
}
\details{
The properties of the Line. When unset, these fields default to values that match the appearance of new lines created in the Slides editor.
}
