% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{LineProperties}
\alias{LineProperties}
\title{LineProperties Object}
\usage{
LineProperties(
  weight = NULL,
  dashStyle = NULL,
  lineFill = NULL,
  endConnection = NULL,
  startArrow = NULL,
  link = NULL,
  endArrow = NULL,
  startConnection = NULL
)
}
\arguments{
\item{weight}{The thickness of the line}

\item{dashStyle}{The dash style of the line}

\item{lineFill}{The fill of the line}

\item{endConnection}{The connection at the end of the line}

\item{startArrow}{The style of the arrow at the beginning of the line}

\item{link}{The hyperlink destination of the line}

\item{endArrow}{The style of the arrow at the end of the line}

\item{startConnection}{The connection at the beginning of the line}
}
\value{
LineProperties object
}
\description{
LineProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of the Line. When unset, these fields default to values that match the appearance of new lines created in the Slides editor.
}
\concept{LineProperties functions}
