% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Link}
\alias{Link}
\title{Link Object}
\usage{
Link(url = NULL, relativeLink = NULL, pageObjectId = NULL, slideIndex = NULL)
}
\arguments{
\item{url}{(string)
If set, indicates this is a link to the external web page at this URL.}

\item{relativeLink}{(string)
If set, indicates this is a link to a slide in this presentation, addressed by its position.}

\item{pageObjectId}{(string)
If set, indicates this is a link to the specific page in this presentation with this ID. A page with this ID may not exist.}

\item{slideIndex}{(integer)
If set, indicates this is a link to the slide at this zero-based index in the presentation. There may not be a slide at this index.}
}
\value{
Link object
}
\description{
Link Object
}
\details{
A hypertext link.
}
