% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Link}
\alias{Link}
\title{Link Object}
\usage{
Link(pageObjectId = NULL, slideIndex = NULL, url = NULL, relativeLink = NULL)
}
\arguments{
\item{pageObjectId}{If set, indicates this is a link to the specific page in this presentation with this ID}

\item{slideIndex}{If set, indicates this is a link to the slide at this zero-based index in the presentation}

\item{url}{If set, indicates this is a link to the external web page at this URL}

\item{relativeLink}{If set, indicates this is a link to a slide in this presentation, addressed by its position}
}
\value{
Link object
}
\description{
Link Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A hypertext link.
}
\concept{Link functions}
