% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{MasterProperties}
\alias{MasterProperties}
\title{MasterProperties Object}
\usage{
MasterProperties(displayName = NULL)
}
\arguments{
\item{displayName}{The human-readable name of the master}
}
\value{
MasterProperties object
}
\description{
MasterProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of Page that are only relevant for pages with page_type MASTER.
}
\concept{MasterProperties functions}
