% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{MergeTableCellsRequest}
\alias{MergeTableCellsRequest}
\title{MergeTableCellsRequest Object}
\usage{
MergeTableCellsRequest(objectId = NULL, tableRange = NULL)
}
\arguments{
\item{objectId}{The object ID of the table}

\item{tableRange}{The table range specifying which cells of the table to merge}
}
\value{
MergeTableCellsRequest object
}
\description{
MergeTableCellsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Merges cells in a Table.
}
\concept{MergeTableCellsRequest functions}
