% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{NotesProperties}
\alias{NotesProperties}
\title{NotesProperties Object}
\usage{
NotesProperties(speakerNotesObjectId = NULL)
}
\arguments{
\item{speakerNotesObjectId}{(string)
The object ID of the shape on this notes page that contains the speaker notes for the corresponding slide. The actual shape may not always exist on the notes page. Inserting text using this object ID will automatically create the shape. In this case, the actual shape may have different object ID. The \code{GetPresentation} or \code{GetPage} action will always return the latest object ID.}
}
\value{
NotesProperties object
}
\description{
NotesProperties Object
}
\details{
The properties of Page that are only relevant for pages with page_type NOTES.
}
