% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Outline}
\alias{Outline}
\title{Outline Object}
\usage{
Outline(
  outlineFill = NULL,
  weight = NULL,
  dashStyle = NULL,
  propertyState = NULL
)
}
\arguments{
\item{outlineFill}{(\link{OutlineFill})
The fill of the outline.}

\item{weight}{(\link{Dimension})
The thickness of the outline.}

\item{dashStyle}{(string)
The dash style of the outline.}

\item{propertyState}{(string)
The outline property state. Updating the outline on a page element will implicitly update this field to \code{RENDERED}, unless another value is specified in the same request. To have no outline on a page element, set this field to \code{NOT_RENDERED}. In this case, any other outline fields set in the same request will be ignored.}
}
\value{
Outline object
}
\description{
Outline Object
}
\details{
The outline of a PageElement. If these fields are unset, they may be inherited from a parent placeholder if it exists. If there is no parent, the fields will default to the value used for new page elements created in the Slides editor, which may depend on the page element kind.
}
