% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Outline}
\alias{Outline}
\title{Outline Object}
\usage{
Outline(
  outlineFill = NULL,
  weight = NULL,
  propertyState = NULL,
  dashStyle = NULL
)
}
\arguments{
\item{outlineFill}{The fill of the outline}

\item{weight}{The thickness of the outline}

\item{propertyState}{The outline property state}

\item{dashStyle}{The dash style of the outline}
}
\value{
Outline object
}
\description{
Outline Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The outline of a PageElement. If these fields are unset, they may be inherited from a parent placeholder if it exists. If there is no parent, the fields will default to the value used for new page elements created in the Slides editor, which may depend on the page element kind.
}
\concept{Outline functions}
