% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Page}
\alias{Page}
\title{Page Object}
\usage{
Page(
  masterProperties = NULL,
  notesProperties = NULL,
  objectId = NULL,
  pageType = NULL,
  revisionId = NULL,
  pageElements = NULL,
  layoutProperties = NULL,
  slideProperties = NULL,
  pageProperties = NULL
)
}
\arguments{
\item{masterProperties}{Master specific properties}

\item{notesProperties}{Notes specific properties}

\item{objectId}{The object ID for this page}

\item{pageType}{The type of the page}

\item{revisionId}{Output only}

\item{pageElements}{The page elements rendered on the page}

\item{layoutProperties}{Layout specific properties}

\item{slideProperties}{Slide specific properties}

\item{pageProperties}{The properties of the page}
}
\value{
Page object
}
\description{
Page Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A page in a presentation.
}
\concept{Page functions}
