% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{PageBackgroundFill}
\alias{PageBackgroundFill}
\title{PageBackgroundFill Object}
\usage{
PageBackgroundFill(
  propertyState = NULL,
  solidFill = NULL,
  stretchedPictureFill = NULL
)
}
\arguments{
\item{propertyState}{(string)
The background fill property state. Updating the fill on a page will implicitly update this field to \code{RENDERED}, unless another value is specified in the same request. To have no fill on a page, set this field to \code{NOT_RENDERED}. In this case, any other fill fields set in the same request will be ignored.}

\item{solidFill}{(\link{SolidFill})
Solid color fill.}

\item{stretchedPictureFill}{(\link{StretchedPictureFill})
Stretched picture fill.}
}
\value{
PageBackgroundFill object
}
\description{
PageBackgroundFill Object
}
\details{
The page background fill.
}
