% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{PageElement}
\alias{PageElement}
\title{PageElement Object}
\usage{
PageElement(
  objectId = NULL,
  size = NULL,
  transform = NULL,
  title = NULL,
  description = NULL,
  elementGroup = NULL,
  shape = NULL,
  image = NULL,
  video = NULL,
  line = NULL,
  table = NULL,
  wordArt = NULL,
  sheetsChart = NULL,
  speakerSpotlight = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID for this page element. Object IDs used by google.apps.slides.v1.Page and google.apps.slides.v1.PageElement share the same namespace.}

\item{size}{(\link{Size})
The size of the page element.}

\item{transform}{(\link{AffineTransform})
The transform of the page element. The visual appearance of the page element is determined by its absolute transform. To compute the absolute transform, preconcatenate a page element's transform with the transforms of all of its parent groups. If the page element is not in a group, its absolute transform is the same as the value in this field. The initial transform for the newly created Group is always the identity transform.}

\item{title}{(string)
The title of the page element. Combined with description to display alt text. The field is not supported for Group elements.}

\item{description}{(string)
The description of the page element. Combined with title to display alt text. The field is not supported for Group elements.}

\item{elementGroup}{(\link{Group})
A collection of page elements joined as a single unit.}

\item{shape}{(\link{Shape})
A generic shape.}

\item{image}{(\link{Image})
An image page element.}

\item{video}{(\link{Video})
A video page element.}

\item{line}{(\link{Line})
A line page element.}

\item{table}{(\link{Table})
A table page element.}

\item{wordArt}{(\link{WordArt})
A word art page element.}

\item{sheetsChart}{(\link{SheetsChart})
A linked chart embedded from Google Sheets. Unlinked charts are represented as images.}

\item{speakerSpotlight}{(\link{SpeakerSpotlight})
A Speaker Spotlight.}
}
\value{
PageElement object
}
\description{
PageElement Object
}
\details{
A visual element rendered on a page.
}
