% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{PageElementProperties}
\alias{PageElementProperties}
\title{PageElementProperties Object}
\usage{
PageElementProperties(pageObjectId = NULL, size = NULL, transform = NULL)
}
\arguments{
\item{pageObjectId}{(string)
The object ID of the page where the element is located.}

\item{size}{(\link{Size})
The size of the element.}

\item{transform}{(\link{AffineTransform})
The transform for the element.}
}
\value{
PageElementProperties object
}
\description{
PageElementProperties Object
}
\details{
Common properties for a page element. Note: When you initially create a PageElement, the API may modify the values of both \code{size} and \code{transform}, but the visual size will be unchanged.
}
