% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ParagraphMarker}
\alias{ParagraphMarker}
\title{ParagraphMarker Object}
\usage{
ParagraphMarker(style = NULL, bullet = NULL)
}
\arguments{
\item{style}{(\link{ParagraphStyle})
The paragraph's style}

\item{bullet}{(\link{Bullet})
The bullet for this paragraph. If not present, the paragraph does not belong to a list.}
}
\value{
ParagraphMarker object
}
\description{
ParagraphMarker Object
}
\details{
A TextElement kind that represents the beginning of a new paragraph.
}
