% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ParagraphStyle}
\alias{ParagraphStyle}
\title{ParagraphStyle Object}
\usage{
ParagraphStyle(
  lineSpacing = NULL,
  indentStart = NULL,
  indentEnd = NULL,
  alignment = NULL,
  direction = NULL,
  indentFirstLine = NULL,
  spacingMode = NULL,
  spaceAbove = NULL,
  spaceBelow = NULL
)
}
\arguments{
\item{lineSpacing}{The amount of space between lines, as a percentage of normal, where normal is represented as 100}

\item{indentStart}{The amount indentation for the paragraph on the side that corresponds to the start of the text, based on the current text direction}

\item{indentEnd}{The amount indentation for the paragraph on the side that corresponds to the end of the text, based on the current text direction}

\item{alignment}{The text alignment for this paragraph}

\item{direction}{The text direction of this paragraph}

\item{indentFirstLine}{The amount of indentation for the start of the first line of the paragraph}

\item{spacingMode}{The spacing mode for the paragraph}

\item{spaceAbove}{The amount of extra space above the paragraph}

\item{spaceBelow}{The amount of extra space below the paragraph}
}
\value{
ParagraphStyle object
}
\description{
ParagraphStyle Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Styles that apply to a whole paragraph. If this text is contained in a shape with a parent placeholder, then these paragraph styles may be inherited from the parent. Which paragraph styles are inherited depend on the nesting level of lists: * A paragraph not in a list will inherit its paragraph style from the paragraph at the 0 nesting level of the list inside the parent placeholder. * A paragraph in a list will inherit its paragraph style from the paragraph at its corresponding nesting level of the list inside the parent placeholder. Inherited paragraph styles are represented as unset fields in this message.
}
\concept{ParagraphStyle functions}
