% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Placeholder}
\alias{Placeholder}
\title{Placeholder Object}
\usage{
Placeholder(type = NULL, index = NULL, parentObjectId = NULL)
}
\arguments{
\item{type}{(string)
The type of the placeholder.
\itemize{
\item \code{NONE}: Default value, signifies it is not a placeholder.
\item \code{BODY}: Body text.
\item \code{CHART}: Chart or graph.
\item \code{CLIP_ART}: Clip art image.
\item \code{CENTERED_TITLE}: Title centered.
\item \code{DIAGRAM}: Diagram.
\item \code{DATE_AND_TIME}: Date and time.
\item \code{FOOTER}: Footer text.
\item \code{HEADER}: Header text.
\item \code{MEDIA}: Multimedia.
\item \code{OBJECT}: Any content type.
\item \code{PICTURE}: Picture.
\item \code{SLIDE_NUMBER}: Number of a slide.
\item \code{SUBTITLE}: Subtitle.
\item \code{TABLE}: Table.
\item \code{TITLE}: Slide title.
\item \code{SLIDE_IMAGE}: Slide image.
}}

\item{index}{(integer)
The index of the placeholder. If the same placeholder types are present in the same page, they would have different index values.}

\item{parentObjectId}{(string)
The object ID of this shape's parent placeholder. If unset, the parent placeholder shape does not exist, so the shape does not inherit properties from any other shape.}
}
\value{
Placeholder object
}
\description{
Placeholder Object
}
\details{
The placeholder information that uniquely identifies a placeholder shape.
}
