% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Placeholder}
\alias{Placeholder}
\title{Placeholder Object}
\usage{
Placeholder(parentObjectId = NULL, index = NULL, type = NULL)
}
\arguments{
\item{parentObjectId}{The object ID of this shape's parent placeholder}

\item{index}{The index of the placeholder}

\item{type}{The type of the placeholder}
}
\value{
Placeholder object
}
\description{
Placeholder Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The placeholder information that uniquely identifies a placeholder shape.
}
\concept{Placeholder functions}
