% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Range}
\alias{Range}
\title{Range Object}
\usage{
Range(startIndex = NULL, endIndex = NULL, type = NULL)
}
\arguments{
\item{startIndex}{(integer)
The optional zero-based index of the beginning of the collection. Required for \code{FIXED_RANGE} and \code{FROM_START_INDEX} ranges.}

\item{endIndex}{(integer)
The optional zero-based index of the end of the collection. Required for \code{FIXED_RANGE} ranges.}

\item{type}{(string)
The type of range.}
}
\value{
Range object
}
\description{
Range Object
}
\details{
Specifies a contiguous range of an indexed collection, such as characters in text.
}
