% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Range}
\alias{Range}
\title{Range Object}
\usage{
Range(startIndex = NULL, type = NULL, endIndex = NULL)
}
\arguments{
\item{startIndex}{The optional zero-based index of the beginning of the collection}

\item{type}{The type of range}

\item{endIndex}{The optional zero-based index of the end of the collection}
}
\value{
Range object
}
\description{
Range Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies a contiguous range of an indexed collection, such as characters in text.
}
\concept{Range functions}
