% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Recolor}
\alias{Recolor}
\title{Recolor Object}
\usage{
Recolor(name = NULL, recolorStops = NULL)
}
\arguments{
\item{name}{The name of the recolor effect}

\item{recolorStops}{The recolor effect is represented by a gradient, which is a list of color stops}
}
\value{
Recolor object
}
\description{
Recolor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A recolor effect applied on an image.
}
\concept{Recolor functions}
