% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ReplaceAllShapesWithImageRequest}
\alias{ReplaceAllShapesWithImageRequest}
\title{ReplaceAllShapesWithImageRequest Object}
\usage{
ReplaceAllShapesWithImageRequest(
  containsText = NULL,
  imageUrl = NULL,
  replaceMethod = NULL,
  imageReplaceMethod = NULL,
  pageObjectIds = NULL
)
}
\arguments{
\item{containsText}{(\link{SubstringMatchCriteria})
If set, this request will replace all of the shapes that contain the given text.}

\item{imageUrl}{(string)
The image URL. The image is fetched once at insertion time and a copy is stored for display inside the presentation. Images must be less than 50MB in size, cannot exceed 25 megapixels, and must be in one of PNG, JPEG, or GIF format. The provided URL can be at most 2 kB in length. The URL itself is saved with the image, and exposed via the Image.source_url field.}

\item{replaceMethod}{(string)
The replace method. \emph{Deprecated}: use \code{image_replace_method} instead. If you specify both a \code{replace_method} and an \code{image_replace_method}, the \code{image_replace_method} takes precedence.}

\item{imageReplaceMethod}{(string)
The image replace method. If you specify both a \code{replace_method} and an \code{image_replace_method}, the \code{image_replace_method} takes precedence. If you do not specify a value for \code{image_replace_method}, but specify a value for \code{replace_method}, then the specified \code{replace_method} value is used. If you do not specify either, then CENTER_INSIDE is used.}

\item{pageObjectIds}{(list of \link{string} objects)
If non-empty, limits the matches to page elements only on the given pages. Returns a 400 bad request error if given the page object ID of a notes page or a notes master, or if a page with that object ID doesn't exist in the presentation.}
}
\value{
ReplaceAllShapesWithImageRequest object
}
\description{
ReplaceAllShapesWithImageRequest Object
}
\details{
Replaces all shapes that match the given criteria with the provided image. The images replacing the shapes are rectangular after being inserted into the presentation and do not take on the forms of the shapes.
}
