% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ReplaceAllShapesWithSheetsChartRequest}
\alias{ReplaceAllShapesWithSheetsChartRequest}
\title{ReplaceAllShapesWithSheetsChartRequest Object}
\usage{
ReplaceAllShapesWithSheetsChartRequest(
  containsText = NULL,
  spreadsheetId = NULL,
  chartId = NULL,
  linkingMode = NULL,
  pageObjectIds = NULL
)
}
\arguments{
\item{containsText}{(\link{SubstringMatchCriteria})
The criteria that the shapes must match in order to be replaced. The request will replace all of the shapes that contain the given text.}

\item{spreadsheetId}{(string)
The ID of the Google Sheets spreadsheet that contains the chart.}

\item{chartId}{(integer)
The ID of the specific chart in the Google Sheets spreadsheet.}

\item{linkingMode}{(string)
The mode with which the chart is linked to the source spreadsheet. When not specified, the chart will be an image that is not linked.
\itemize{
\item \code{NOT_LINKED_IMAGE}: The chart is not associated with the source spreadsheet and cannot be updated. A chart that is not linked will be inserted as an image.
\item \code{LINKED}: Linking the chart allows it to be updated, and other collaborators will see a link to the spreadsheet.
}}

\item{pageObjectIds}{(list of \link{string} objects)
If non-empty, limits the matches to page elements only on the given pages. Returns a 400 bad request error if given the page object ID of a notes page or a notes master, or if a page with that object ID doesn't exist in the presentation.}
}
\value{
ReplaceAllShapesWithSheetsChartRequest object
}
\description{
ReplaceAllShapesWithSheetsChartRequest Object
}
\details{
Replaces all shapes that match the given criteria with the provided Google Sheets chart. The chart will be scaled and centered to fit within the bounds of the original shape. NOTE: Replacing shapes with a chart requires at least one of the spreadsheets.readonly, spreadsheets, drive.readonly, or drive OAuth scopes.
}
