% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ReplaceImageRequest}
\alias{ReplaceImageRequest}
\title{ReplaceImageRequest Object}
\usage{
ReplaceImageRequest(
  imageObjectId = NULL,
  url = NULL,
  imageReplaceMethod = NULL
)
}
\arguments{
\item{imageObjectId}{(string)
The ID of the existing image that will be replaced. The ID can be retrieved from the response of a get request.}

\item{url}{(string)
The image URL. The image is fetched once at insertion time and a copy is stored for display inside the presentation. Images must be less than 50MB, cannot exceed 25 megapixels, and must be in PNG, JPEG, or GIF format. The provided URL can't surpass 2 KB in length. The URL is saved with the image, and exposed through the Image.source_url field.}

\item{imageReplaceMethod}{(string)
The replacement method.
\itemize{
\item \code{IMAGE_REPLACE_METHOD_UNSPECIFIED}: Unspecified image replace method. This value must not be used.
\item \code{CENTER_INSIDE}: Scales and centers the image to fit within the bounds of the original shape and maintains the image's aspect ratio. The rendered size of the image may be smaller than the size of the shape. This is the default method when one is not specified.
\item \code{CENTER_CROP}: Scales and centers the image to fill the bounds of the original shape. The image may be cropped in order to fill the shape. The rendered size of the image will be the same as the original shape.
}}
}
\value{
ReplaceImageRequest object
}
\description{
ReplaceImageRequest Object
}
\details{
Replaces an existing image with a new image. Replacing an image removes some image effects from the existing image.
}
