% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Request}
\alias{Request}
\title{Request Object}
\usage{
Request(...)
}
\arguments{
\item{createSlide}{(\link{CreateSlideRequest})
Creates a new slide.}

\item{createShape}{(\link{CreateShapeRequest})
Creates a new shape.}

\item{createTable}{(\link{CreateTableRequest})
Creates a new table.}

\item{insertText}{(\link{InsertTextRequest})
Inserts text into a shape or table cell.}

\item{insertTableRows}{(\link{InsertTableRowsRequest})
Inserts rows into a table.}

\item{insertTableColumns}{(\link{InsertTableColumnsRequest})
Inserts columns into a table.}

\item{deleteTableRow}{(\link{DeleteTableRowRequest})
Deletes a row from a table.}

\item{deleteTableColumn}{(\link{DeleteTableColumnRequest})
Deletes a column from a table.}

\item{replaceAllText}{(\link{ReplaceAllTextRequest})
Replaces all instances of specified text.}

\item{deleteObject}{(\link{DeleteObjectRequest})
Deletes a page or page element from the presentation.}

\item{updatePageElementTransform}{(\link{UpdatePageElementTransformRequest})
Updates the transform of a page element.}

\item{updateSlidesPosition}{(\link{UpdateSlidesPositionRequest})
Updates the position of a set of slides in the presentation.}

\item{deleteText}{(\link{DeleteTextRequest})
Deletes text from a shape or a table cell.}

\item{createImage}{(\link{CreateImageRequest})
Creates an image.}

\item{createVideo}{(\link{CreateVideoRequest})
Creates a video.}

\item{createSheetsChart}{(\link{CreateSheetsChartRequest})
Creates an embedded Google Sheets chart.}

\item{createLine}{(\link{CreateLineRequest})
Creates a line.}

\item{refreshSheetsChart}{(\link{RefreshSheetsChartRequest})
Refreshes a Google Sheets chart.}

\item{updateShapeProperties}{(\link{UpdateShapePropertiesRequest})
Updates the properties of a Shape.}

\item{updateImageProperties}{(\link{UpdateImagePropertiesRequest})
Updates the properties of an Image.}

\item{updateVideoProperties}{(\link{UpdateVideoPropertiesRequest})
Updates the properties of a Video.}

\item{updatePageProperties}{(\link{UpdatePagePropertiesRequest})
Updates the properties of a Page.}

\item{updateTableCellProperties}{(\link{UpdateTableCellPropertiesRequest})
Updates the properties of a TableCell.}

\item{updateLineProperties}{(\link{UpdateLinePropertiesRequest})
Updates the properties of a Line.}

\item{createParagraphBullets}{(\link{CreateParagraphBulletsRequest})
Creates bullets for paragraphs.}

\item{replaceAllShapesWithImage}{(\link{ReplaceAllShapesWithImageRequest})
Replaces all shapes matching some criteria with an image.}

\item{duplicateObject}{(\link{DuplicateObjectRequest})
Duplicates a slide or page element.}

\item{updateTextStyle}{(\link{UpdateTextStyleRequest})
Updates the styling of text within a Shape or Table.}

\item{replaceAllShapesWithSheetsChart}{(\link{ReplaceAllShapesWithSheetsChartRequest})
Replaces all shapes matching some criteria with a Google Sheets chart.}

\item{deleteParagraphBullets}{(\link{DeleteParagraphBulletsRequest})
Deletes bullets from paragraphs.}

\item{updateParagraphStyle}{(\link{UpdateParagraphStyleRequest})
Updates the styling of paragraphs within a Shape or Table.}

\item{updateTableBorderProperties}{(\link{UpdateTableBorderPropertiesRequest})
Updates the properties of the table borders in a Table.}

\item{updateTableColumnProperties}{(\link{UpdateTableColumnPropertiesRequest})
Updates the properties of a Table column.}

\item{updateTableRowProperties}{(\link{UpdateTableRowPropertiesRequest})
Updates the properties of a Table row.}

\item{mergeTableCells}{(\link{MergeTableCellsRequest})
Merges cells in a Table.}

\item{unmergeTableCells}{(\link{UnmergeTableCellsRequest})
Unmerges cells in a Table.}

\item{groupObjects}{(\link{GroupObjectsRequest})
Groups objects, such as page elements.}

\item{ungroupObjects}{(\link{UngroupObjectsRequest})
Ungroups objects, such as groups.}

\item{updatePageElementAltText}{(\link{UpdatePageElementAltTextRequest})
Updates the alt text title and/or description of a page element.}

\item{replaceImage}{(\link{ReplaceImageRequest})
Replaces an existing image with a new image.}

\item{updateSlideProperties}{(\link{UpdateSlidePropertiesRequest})
Updates the properties of a Slide}

\item{updatePageElementsZOrder}{(\link{UpdatePageElementsZOrderRequest})
Updates the Z-order of page elements.}

\item{updateLineCategory}{(\link{UpdateLineCategoryRequest})
Updates the category of a line.}

\item{rerouteLine}{(\link{RerouteLineRequest})
Reroutes a line such that it's connected at the two closest connection sites on the connected page elements.}
}
\value{
Request object
}
\description{
Request Object
}
\details{
A single kind of update to apply to a presentation.
}
