% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{RerouteLineRequest}
\alias{RerouteLineRequest}
\title{RerouteLineRequest Object}
\usage{
RerouteLineRequest(objectId = NULL)
}
\arguments{
\item{objectId}{(string)
The object ID of the line to reroute. Only a line with a category indicating it is a "connector" can be rerouted. The start and end connections of the line must be on different page elements.}
}
\value{
RerouteLineRequest object
}
\description{
RerouteLineRequest Object
}
\details{
Reroutes a line such that it's connected at the two closest connection sites on the connected page elements.
}
