% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Response}
\alias{Response}
\title{Response Object}
\usage{
Response(
  createLine = NULL,
  createImage = NULL,
  createSheetsChart = NULL,
  createSlide = NULL,
  replaceAllText = NULL,
  createShape = NULL,
  createVideo = NULL,
  replaceAllShapesWithImage = NULL,
  duplicateObject = NULL,
  createTable = NULL,
  groupObjects = NULL,
  replaceAllShapesWithSheetsChart = NULL
)
}
\arguments{
\item{createLine}{The result of creating a line}

\item{createImage}{The result of creating an image}

\item{createSheetsChart}{The result of creating a Google Sheets chart}

\item{createSlide}{The result of creating a slide}

\item{replaceAllText}{The result of replacing text}

\item{createShape}{The result of creating a shape}

\item{createVideo}{The result of creating a video}

\item{replaceAllShapesWithImage}{The result of replacing all shapes matching some criteria with an image}

\item{duplicateObject}{The result of duplicating an object}

\item{createTable}{The result of creating a table}

\item{groupObjects}{The result of grouping objects}

\item{replaceAllShapesWithSheetsChart}{The result of replacing all shapes matching some criteria with a Google Sheets chart}
}
\value{
Response object
}
\description{
Response Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single response from an update.
}
\concept{Response functions}
