% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_null_objs}
\alias{rm_null_objs}
\title{Recursively step down into list, removing all such objects}
\usage{
rm_null_objs(x)
}
\description{
Recursively step down into list, removing all such objects
}
\keyword{internal}
