% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Shadow}
\alias{Shadow}
\title{Shadow Object}
\usage{
Shadow(
  transform = NULL,
  propertyState = NULL,
  blurRadius = NULL,
  type = NULL,
  alpha = NULL,
  rotateWithShape = NULL,
  alignment = NULL,
  color = NULL
)
}
\arguments{
\item{transform}{Transform that encodes the translate, scale, and skew of the shadow, relative to the alignment position}

\item{propertyState}{The shadow property state}

\item{blurRadius}{The radius of the shadow blur}

\item{type}{The type of the shadow}

\item{alpha}{The alpha of the shadow's color, from 0}

\item{rotateWithShape}{Whether the shadow should rotate with the shape}

\item{alignment}{The alignment point of the shadow, that sets the origin for translate, scale and skew of the shadow}

\item{color}{The shadow color value}
}
\value{
Shadow object
}
\description{
Shadow Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The shadow properties of a page element. If these fields are unset, they may be inherited from a parent placeholder if it exists. If there is no parent, the fields will default to the value used for new page elements created in the Slides editor, which may depend on the page element kind.
}
\concept{Shadow functions}
