% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Shape}
\alias{Shape}
\title{Shape Object}
\usage{
Shape(
  shapeProperties = NULL,
  shapeType = NULL,
  text = NULL,
  placeholder = NULL
)
}
\arguments{
\item{shapeProperties}{The properties of the shape}

\item{shapeType}{The type of the shape}

\item{text}{The text content of the shape}

\item{placeholder}{Placeholders are page elements that inherit from corresponding placeholders on layouts and masters}
}
\value{
Shape object
}
\description{
Shape Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A PageElement kind representing a generic shape that does not have a more specific classification.
}
\concept{Shape functions}
