% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{ShapeProperties}
\alias{ShapeProperties}
\title{ShapeProperties Object}
\usage{
ShapeProperties(
  shapeBackgroundFill = NULL,
  outline = NULL,
  shadow = NULL,
  link = NULL,
  contentAlignment = NULL,
  autofit = NULL
)
}
\arguments{
\item{shapeBackgroundFill}{(\link{ShapeBackgroundFill})
The background fill of the shape. If unset, the background fill is inherited from a parent placeholder if it exists. If the shape has no parent, then the default background fill depends on the shape type, matching the defaults for new shapes created in the Slides editor.}

\item{outline}{(\link{Outline})
The outline of the shape. If unset, the outline is inherited from a parent placeholder if it exists. If the shape has no parent, then the default outline depends on the shape type, matching the defaults for new shapes created in the Slides editor.}

\item{shadow}{(\link{Shadow})
The shadow properties of the shape. If unset, the shadow is inherited from a parent placeholder if it exists. If the shape has no parent, then the default shadow matches the defaults for new shapes created in the Slides editor. This property is read-only.}

\item{link}{(\link{Link})
The hyperlink destination of the shape. If unset, there is no link. Links are not inherited from parent placeholders.}

\item{contentAlignment}{(string)
The alignment of the content in the shape. If unspecified, the alignment is inherited from a parent placeholder if it exists. If the shape has no parent, the default alignment matches the alignment for new shapes created in the Slides editor.
\itemize{
\item \code{CONTENT_ALIGNMENT_UNSPECIFIED}: An unspecified content alignment. The content alignment is inherited from the parent if it exists.
\item \code{CONTENT_ALIGNMENT_UNSUPPORTED}: An unsupported content alignment.
\item \code{TOP}: An alignment that aligns the content to the top of the content holder. Corresponds to ECMA-376 ST_TextAnchoringType 't'.
\item \code{MIDDLE}: An alignment that aligns the content to the middle of the content holder. Corresponds to ECMA-376 ST_TextAnchoringType 'ctr'.
\item \code{BOTTOM}: An alignment that aligns the content to the bottom of the content holder. Corresponds to ECMA-376 ST_TextAnchoringType 'b'.
}}

\item{autofit}{(\link{Autofit})
The autofit properties of the shape. This property is only set for shapes that allow text.}
}
\value{
ShapeProperties object
}
\description{
ShapeProperties Object
}
\details{
The properties of a Shape. If the shape is a placeholder shape as determined by the placeholder field, then these properties may be inherited from a parent placeholder shape. Determining the rendered value of the property depends on the corresponding property_state field value. Any text autofit settings on the shape are automatically deactivated by requests that can impact how text fits in the shape.
}
