% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{SheetsChart}
\alias{SheetsChart}
\title{SheetsChart Object}
\usage{
SheetsChart(
  spreadsheetId = NULL,
  chartId = NULL,
  contentUrl = NULL,
  sheetsChartProperties = NULL
)
}
\arguments{
\item{spreadsheetId}{(string)
The ID of the Google Sheets spreadsheet that contains the source chart.}

\item{chartId}{(integer)
The ID of the specific chart in the Google Sheets spreadsheet that is embedded.}

\item{contentUrl}{(string)
The URL of an image of the embedded chart, with a default lifetime of 30 minutes. This URL is tagged with the account of the requester. Anyone with the URL effectively accesses the image as the original requester. Access to the image may be lost if the presentation's sharing settings change.}

\item{sheetsChartProperties}{(\link{SheetsChartProperties})
The properties of the Sheets chart.}
}
\value{
SheetsChart object
}
\description{
SheetsChart Object
}
\details{
A PageElement kind representing a linked chart embedded from Google Sheets.
}
