% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Size}
\alias{Size}
\title{Size Object}
\usage{
Size(width = NULL, height = NULL)
}
\arguments{
\item{width}{The width of the object}

\item{height}{The height of the object}
}
\value{
Size object
}
\description{
Size Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A width and height.
}
\concept{Size functions}
