% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{SlideProperties}
\alias{SlideProperties}
\title{SlideProperties Object}
\usage{
SlideProperties(
  layoutObjectId = NULL,
  masterObjectId = NULL,
  notesPage = NULL,
  isSkipped = NULL
)
}
\arguments{
\item{layoutObjectId}{(string)
The object ID of the layout that this slide is based on. This property is read-only.}

\item{masterObjectId}{(string)
The object ID of the master that this slide is based on. This property is read-only.}

\item{notesPage}{(\link{Page})
The notes page that this slide is associated with. It defines the visual appearance of a notes page when printing or exporting slides with speaker notes. A notes page inherits properties from the notes master. The placeholder shape with type BODY on the notes page contains the speaker notes for this slide. The ID of this shape is identified by the speakerNotesObjectId field. The notes page is read-only except for the text content and styles of the speaker notes shape. This property is read-only.}

\item{isSkipped}{(boolean)
Whether the slide is skipped in the presentation mode. Defaults to false.}
}
\value{
SlideProperties object
}
\description{
SlideProperties Object
}
\details{
The properties of Page that are only relevant for pages with page_type SLIDE.
}
