% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableBorderCell}
\alias{TableBorderCell}
\title{TableBorderCell Object}
\usage{
TableBorderCell(location = NULL, tableBorderProperties = NULL)
}
\arguments{
\item{location}{The location of the border within the border table}

\item{tableBorderProperties}{The border properties}
}
\value{
TableBorderCell object
}
\description{
TableBorderCell Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of each border cell.
}
\concept{TableBorderCell functions}
