% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableBorderFill}
\alias{TableBorderFill}
\title{TableBorderFill Object}
\usage{
TableBorderFill(solidFill = NULL)
}
\arguments{
\item{solidFill}{Solid fill}
}
\value{
TableBorderFill object
}
\description{
TableBorderFill Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The fill of the border.
}
\concept{TableBorderFill functions}
