% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableBorderProperties}
\alias{TableBorderProperties}
\title{TableBorderProperties Object}
\usage{
TableBorderProperties(dashStyle = NULL, weight = NULL, tableBorderFill = NULL)
}
\arguments{
\item{dashStyle}{The dash style of the border}

\item{weight}{The thickness of the border}

\item{tableBorderFill}{The fill of the table border}
}
\value{
TableBorderProperties object
}
\description{
TableBorderProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The border styling properties of the TableBorderCell.
}
\concept{TableBorderProperties functions}
