% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableBorderRow}
\alias{TableBorderRow}
\title{TableBorderRow Object}
\usage{
TableBorderRow(tableBorderCells = NULL)
}
\arguments{
\item{tableBorderCells}{(list of \link{TableBorderCell} objects)
Properties of each border cell. When a border's adjacent table cells are merged, it is not included in the response.}
}
\value{
TableBorderRow object
}
\description{
TableBorderRow Object
}
\details{
Contents of each border row in a table.
}
