% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableCellBackgroundFill}
\alias{TableCellBackgroundFill}
\title{TableCellBackgroundFill Object}
\usage{
TableCellBackgroundFill(propertyState = NULL, solidFill = NULL)
}
\arguments{
\item{propertyState}{(string)
The background fill property state. Updating the fill on a table cell will implicitly update this field to \code{RENDERED}, unless another value is specified in the same request. To have no fill on a table cell, set this field to \code{NOT_RENDERED}. In this case, any other fill fields set in the same request will be ignored.}

\item{solidFill}{(\link{SolidFill})
Solid color fill.}
}
\value{
TableCellBackgroundFill object
}
\description{
TableCellBackgroundFill Object
}
\details{
The table cell background fill.
}
