% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableCellLocation}
\alias{TableCellLocation}
\title{TableCellLocation Object}
\usage{
TableCellLocation(rowIndex = NULL, columnIndex = NULL)
}
\arguments{
\item{rowIndex}{(integer)
The 0-based row index.}

\item{columnIndex}{(integer)
The 0-based column index.}
}
\value{
TableCellLocation object
}
\description{
TableCellLocation Object
}
\details{
A location of a single table cell within a table.
}
