% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableCellProperties}
\alias{TableCellProperties}
\title{TableCellProperties Object}
\usage{
TableCellProperties(tableCellBackgroundFill = NULL, contentAlignment = NULL)
}
\arguments{
\item{tableCellBackgroundFill}{The background fill of the table cell}

\item{contentAlignment}{The alignment of the content in the table cell}
}
\value{
TableCellProperties object
}
\description{
TableCellProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of the TableCell.
}
\concept{TableCellProperties functions}
