% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableCellProperties}
\alias{TableCellProperties}
\title{TableCellProperties Object}
\usage{
TableCellProperties(tableCellBackgroundFill = NULL, contentAlignment = NULL)
}
\arguments{
\item{tableCellBackgroundFill}{(\link{TableCellBackgroundFill})
The background fill of the table cell. The default fill matches the fill for newly created table cells in the Slides editor.}

\item{contentAlignment}{(string)
The alignment of the content in the table cell. The default alignment matches the alignment for newly created table cells in the Slides editor.
\itemize{
\item \code{CONTENT_ALIGNMENT_UNSPECIFIED}: An unspecified content alignment. The content alignment is inherited from the parent if it exists.
\item \code{CONTENT_ALIGNMENT_UNSUPPORTED}: An unsupported content alignment.
\item \code{TOP}: An alignment that aligns the content to the top of the content holder. Corresponds to ECMA-376 ST_TextAnchoringType 't'.
\item \code{MIDDLE}: An alignment that aligns the content to the middle of the content holder. Corresponds to ECMA-376 ST_TextAnchoringType 'ctr'.
\item \code{BOTTOM}: An alignment that aligns the content to the bottom of the content holder. Corresponds to ECMA-376 ST_TextAnchoringType 'b'.
}}
}
\value{
TableCellProperties object
}
\description{
TableCellProperties Object
}
\details{
The properties of the TableCell.
}
