% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TableRowProperties}
\alias{TableRowProperties}
\title{TableRowProperties Object}
\usage{
TableRowProperties(minRowHeight = NULL)
}
\arguments{
\item{minRowHeight}{(\link{Dimension})
Minimum height of the row. The row will be rendered in the Slides editor at a height equal to or greater than this value in order to show all the text in the row's cell(s).}
}
\value{
TableRowProperties object
}
\description{
TableRowProperties Object
}
\details{
Properties of each row in a table.
}
