% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TextElement}
\alias{TextElement}
\title{TextElement Object}
\usage{
TextElement(
  startIndex = NULL,
  paragraphMarker = NULL,
  textRun = NULL,
  endIndex = NULL,
  autoText = NULL
)
}
\arguments{
\item{startIndex}{The zero-based start index of this text element, in Unicode code units}

\item{paragraphMarker}{A marker representing the beginning of a new paragraph}

\item{textRun}{A TextElement representing a run of text where all of the characters in the run have the same TextStyle}

\item{endIndex}{The zero-based end index of this text element, exclusive, in Unicode code units}

\item{autoText}{A TextElement representing a spot in the text that is dynamically replaced with content that can change over time}
}
\value{
TextElement object
}
\description{
TextElement Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TextElement describes the content of a range of indices in the text content of a Shape or TableCell.
}
\concept{TextElement functions}
