% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{TextStyle}
\alias{TextStyle}
\title{TextStyle Object}
\usage{
TextStyle(
  link = NULL,
  baselineOffset = NULL,
  foregroundColor = NULL,
  strikethrough = NULL,
  weightedFontFamily = NULL,
  underline = NULL,
  italic = NULL,
  smallCaps = NULL,
  fontSize = NULL,
  fontFamily = NULL,
  bold = NULL,
  backgroundColor = NULL
)
}
\arguments{
\item{link}{The hyperlink destination of the text}

\item{baselineOffset}{The text's vertical offset from its normal position}

\item{foregroundColor}{The color of the text itself}

\item{strikethrough}{Whether or not the text is struck through}

\item{weightedFontFamily}{The font family and rendered weight of the text}

\item{underline}{Whether or not the text is underlined}

\item{italic}{Whether or not the text is italicized}

\item{smallCaps}{Whether or not the text is in small capital letters}

\item{fontSize}{The size of the text's font}

\item{fontFamily}{The font family of the text}

\item{bold}{Whether or not the text is rendered as bold}

\item{backgroundColor}{The background color of the text}
}
\value{
TextStyle object
}
\description{
TextStyle Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the styling that can be applied to a TextRun. If this text is contained in a shape with a parent placeholder, then these text styles may be inherited from the parent. Which text styles are inherited depend on the nesting level of lists: * A text run in a paragraph that is not in a list will inherit its text style from the the newline character in the paragraph at the 0 nesting level of the list inside the parent placeholder. * A text run in a paragraph that is in a list will inherit its text style from the newline character in the paragraph at its corresponding nesting level of the list inside the parent placeholder. Inherited text styles are represented as unset fields in this message. If text is contained in a shape without a parent placeholder, unsetting these fields will revert the style to a value matching the defaults in the Slides editor.
}
\concept{TextStyle functions}
