% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Thumbnail}
\alias{Thumbnail}
\title{Thumbnail Object}
\usage{
Thumbnail(width = NULL, height = NULL, contentUrl = NULL)
}
\arguments{
\item{width}{(integer)
The positive width in pixels of the thumbnail image.}

\item{height}{(integer)
The positive height in pixels of the thumbnail image.}

\item{contentUrl}{(string)
The content URL of the thumbnail image. The URL to the image has a default lifetime of 30 minutes. This URL is tagged with the account of the requester. Anyone with the URL effectively accesses the image as the original requester. Access to the image may be lost if the presentation's sharing settings change. The mime type of the thumbnail image is the same as specified in the \code{GetPageThumbnailRequest}.}
}
\value{
Thumbnail object
}
\description{
Thumbnail Object
}
\details{
The thumbnail of a page.
}
