% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UngroupObjectsRequest}
\alias{UngroupObjectsRequest}
\title{UngroupObjectsRequest Object}
\usage{
UngroupObjectsRequest(objectIds = NULL)
}
\arguments{
\item{objectIds}{(list of \link{string} objects)
The object IDs of the objects to ungroup. Only groups that are not inside other groups can be ungrouped. All the groups should be on the same page. The group itself is deleted. The visual sizes and positions of all the children are preserved.}
}
\value{
UngroupObjectsRequest object
}
\description{
UngroupObjectsRequest Object
}
\details{
Ungroups objects, such as groups.
}
