% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UnmergeTableCellsRequest}
\alias{UnmergeTableCellsRequest}
\title{UnmergeTableCellsRequest Object}
\usage{
UnmergeTableCellsRequest(tableRange = NULL, objectId = NULL)
}
\arguments{
\item{tableRange}{The table range specifying which cells of the table to unmerge}

\item{objectId}{The object ID of the table}
}
\value{
UnmergeTableCellsRequest object
}
\description{
UnmergeTableCellsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Unmerges cells in a Table.
}
\concept{UnmergeTableCellsRequest functions}
