% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateLineCategoryRequest}
\alias{UpdateLineCategoryRequest}
\title{UpdateLineCategoryRequest Object}
\usage{
UpdateLineCategoryRequest(lineCategory = NULL, objectId = NULL)
}
\arguments{
\item{lineCategory}{The line category to update to}

\item{objectId}{The object ID of the line the update is applied to}
}
\value{
UpdateLineCategoryRequest object
}
\description{
UpdateLineCategoryRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the category of a line.
}
\concept{UpdateLineCategoryRequest functions}
