% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateLineCategoryRequest}
\alias{UpdateLineCategoryRequest}
\title{UpdateLineCategoryRequest Object}
\usage{
UpdateLineCategoryRequest(objectId = NULL, lineCategory = NULL)
}
\arguments{
\item{objectId}{(string)
The object ID of the line the update is applied to. Only a line with a category indicating it is a "connector" can be updated. The line may be rerouted after updating its category.}

\item{lineCategory}{(string)
The line category to update to. The exact line type is determined based on the category to update to and how it's routed to connect to other page elements.
\itemize{
\item \code{LINE_CATEGORY_UNSPECIFIED}: Unspecified line category.
\item \code{STRAIGHT}: Straight connectors, including straight connector 1.
\item \code{BENT}: Bent connectors, including bent connector 2 to 5.
\item \code{CURVED}: Curved connectors, including curved connector 2 to 5.
}}
}
\value{
UpdateLineCategoryRequest object
}
\description{
UpdateLineCategoryRequest Object
}
\details{
Updates the category of a line.
}
