% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdatePageElementAltTextRequest}
\alias{UpdatePageElementAltTextRequest}
\title{UpdatePageElementAltTextRequest Object}
\usage{
UpdatePageElementAltTextRequest(
  objectId = NULL,
  description = NULL,
  title = NULL
)
}
\arguments{
\item{objectId}{The object ID of the page element the updates are applied to}

\item{description}{The updated alt text description of the page element}

\item{title}{The updated alt text title of the page element}
}
\value{
UpdatePageElementAltTextRequest object
}
\description{
UpdatePageElementAltTextRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the alt text title and/or description of a page element.
}
\concept{UpdatePageElementAltTextRequest functions}
