% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdatePageElementsZOrderRequest}
\alias{UpdatePageElementsZOrderRequest}
\title{UpdatePageElementsZOrderRequest Object}
\usage{
UpdatePageElementsZOrderRequest(pageElementObjectIds = NULL, operation = NULL)
}
\arguments{
\item{pageElementObjectIds}{(list of \link{string} objects)
The object IDs of the page elements to update. All the page elements must be on the same page and must not be grouped.}

\item{operation}{(string)
The Z-order operation to apply on the page elements. When applying the operation on multiple page elements, the relative Z-orders within these page elements before the operation is maintained.}
}
\value{
UpdatePageElementsZOrderRequest object
}
\description{
UpdatePageElementsZOrderRequest Object
}
\details{
Updates the Z-order of page elements. Z-order is an ordering of the elements on the page from back to front. The page element in the front may cover the elements that are behind it.
}
