% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdatePageElementTransformRequest}
\alias{UpdatePageElementTransformRequest}
\title{UpdatePageElementTransformRequest Object}
\usage{
UpdatePageElementTransformRequest(
  objectId = NULL,
  transform = NULL,
  applyMode = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID of the page element to update.}

\item{transform}{(\link{AffineTransform})
The input transform matrix used to update the page element.}

\item{applyMode}{(string)
The apply mode of the transform update.
\itemize{
\item \code{APPLY_MODE_UNSPECIFIED}: Unspecified mode.
\item \code{RELATIVE}: Applies the new AffineTransform matrix to the existing one, and replaces the existing one with the resulting concatenation.
\item \code{ABSOLUTE}: Replaces the existing AffineTransform matrix with the new one.
}}
}
\value{
UpdatePageElementTransformRequest object
}
\description{
UpdatePageElementTransformRequest Object
}
\details{
Updates the transform of a page element. Updating the transform of a group will change the absolute transform of the page elements in that group, which can change their visual appearance. See the documentation for PageElement.transform for more details.
}
