% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateParagraphStyleRequest}
\alias{UpdateParagraphStyleRequest}
\title{UpdateParagraphStyleRequest Object}
\usage{
UpdateParagraphStyleRequest(
  objectId = NULL,
  cellLocation = NULL,
  style = NULL,
  textRange = NULL,
  fields = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID of the shape or table with the text to be styled.}

\item{cellLocation}{(\link{TableCellLocation})
The location of the cell in the table containing the paragraph(s) to style. If \code{object_id} refers to a table, \code{cell_location} must have a value. Otherwise, it must not.}

\item{style}{(\link{ParagraphStyle})
The paragraph's style.}

\item{textRange}{(\link{Range})
The range of text containing the paragraph(s) to style.}

\item{fields}{(string)
The fields that should be updated. At least one field must be specified. The root \code{style} is implied and should not be specified. A single \code{"*"} can be used as short-hand for listing every field. For example, to update the paragraph alignment, set \code{fields} to \code{"alignment"}. To reset a property to its default value, include its field name in the field mask but leave the field itself unset.}
}
\value{
UpdateParagraphStyleRequest object
}
\description{
UpdateParagraphStyleRequest Object
}
\details{
Updates the styling for all of the paragraphs within a Shape or Table that overlap with the given text index range.
}
