% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateShapePropertiesRequest}
\alias{UpdateShapePropertiesRequest}
\title{UpdateShapePropertiesRequest Object}
\usage{
UpdateShapePropertiesRequest(
  objectId = NULL,
  shapeProperties = NULL,
  fields = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID of the shape the updates are applied to.}

\item{shapeProperties}{(\link{ShapeProperties})
The shape properties to update.}

\item{fields}{(string)
The fields that should be updated. At least one field must be specified. The root \code{shapeProperties} is implied and should not be specified. A single \code{"*"} can be used as short-hand for listing every field. For example to update the shape background solid fill color, set \code{fields} to \code{"shapeBackgroundFill.solidFill.color"}. To reset a property to its default value, include its field name in the field mask but leave the field itself unset.}
}
\value{
UpdateShapePropertiesRequest object
}
\description{
UpdateShapePropertiesRequest Object
}
\details{
Update the properties of a Shape.
}
