% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateSlidePropertiesRequest}
\alias{UpdateSlidePropertiesRequest}
\title{UpdateSlidePropertiesRequest Object}
\usage{
UpdateSlidePropertiesRequest(
  objectId = NULL,
  slideProperties = NULL,
  fields = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID of the slide the update is applied to.}

\item{slideProperties}{(\link{SlideProperties})
The slide properties to update.}

\item{fields}{(string)
The fields that should be updated. At least one field must be specified. The root 'slideProperties' is implied and should not be specified. A single \code{"*"} can be used as short-hand for listing every field. For example to update whether a slide is skipped, set \code{fields} to \code{"isSkipped"}. To reset a property to its default value, include its field name in the field mask but leave the field itself unset.}
}
\value{
UpdateSlidePropertiesRequest object
}
\description{
UpdateSlidePropertiesRequest Object
}
\details{
Updates the properties of a Slide.
}
