% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateTableBorderPropertiesRequest}
\alias{UpdateTableBorderPropertiesRequest}
\title{UpdateTableBorderPropertiesRequest Object}
\usage{
UpdateTableBorderPropertiesRequest(
  borderPosition = NULL,
  objectId = NULL,
  tableBorderProperties = NULL,
  tableRange = NULL,
  fields = NULL
)
}
\arguments{
\item{borderPosition}{The border position in the table range the updates should apply to}

\item{objectId}{The object ID of the table}

\item{tableBorderProperties}{The table border properties to update}

\item{tableRange}{The table range representing the subset of the table to which the updates are applied}

\item{fields}{The fields that should be updated}
}
\value{
UpdateTableBorderPropertiesRequest object
}
\description{
UpdateTableBorderPropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the properties of the table borders in a Table.
}
\concept{UpdateTableBorderPropertiesRequest functions}
