% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateTableCellPropertiesRequest}
\alias{UpdateTableCellPropertiesRequest}
\title{UpdateTableCellPropertiesRequest Object}
\usage{
UpdateTableCellPropertiesRequest(
  tableRange = NULL,
  tableCellProperties = NULL,
  objectId = NULL,
  fields = NULL
)
}
\arguments{
\item{tableRange}{The table range representing the subset of the table to which the updates are applied}

\item{tableCellProperties}{The table cell properties to update}

\item{objectId}{The object ID of the table}

\item{fields}{The fields that should be updated}
}
\value{
UpdateTableCellPropertiesRequest object
}
\description{
UpdateTableCellPropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Update the properties of a TableCell.
}
\concept{UpdateTableCellPropertiesRequest functions}
