% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateTableColumnPropertiesRequest}
\alias{UpdateTableColumnPropertiesRequest}
\title{UpdateTableColumnPropertiesRequest Object}
\usage{
UpdateTableColumnPropertiesRequest(
  objectId = NULL,
  columnIndices = NULL,
  tableColumnProperties = NULL,
  fields = NULL
)
}
\arguments{
\item{objectId}{(string)
The object ID of the table.}

\item{columnIndices}{(list of \link{integer} objects)
The list of zero-based indices specifying which columns to update. If no indices are provided, all columns in the table will be updated.}

\item{tableColumnProperties}{(\link{TableColumnProperties})
The table column properties to update. If the value of \code{table_column_properties#column_width} in the request is less than 406,400 EMU (32 points), a 400 bad request error is returned.}

\item{fields}{(string)
The fields that should be updated. At least one field must be specified. The root \code{tableColumnProperties} is implied and should not be specified. A single \code{"*"} can be used as short-hand for listing every field. For example to update the column width, set \code{fields} to \code{"column_width"}. If '"column_width"' is included in the field mask but the property is left unset, the column width will default to 406,400 EMU (32 points).}
}
\value{
UpdateTableColumnPropertiesRequest object
}
\description{
UpdateTableColumnPropertiesRequest Object
}
\details{
Updates the properties of a Table column.
}
