% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{UpdateTableRowPropertiesRequest}
\alias{UpdateTableRowPropertiesRequest}
\title{UpdateTableRowPropertiesRequest Object}
\usage{
UpdateTableRowPropertiesRequest(
  objectId = NULL,
  fields = NULL,
  tableRowProperties = NULL,
  rowIndices = NULL
)
}
\arguments{
\item{objectId}{The object ID of the table}

\item{fields}{The fields that should be updated}

\item{tableRowProperties}{The table row properties to update}

\item{rowIndices}{The list of zero-based indices specifying which rows to update}
}
\value{
UpdateTableRowPropertiesRequest object
}
\description{
UpdateTableRowPropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the properties of a Table row.
}
\concept{UpdateTableRowPropertiesRequest functions}
