% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Video}
\alias{Video}
\title{Video Object}
\usage{
Video(url = NULL, source = NULL, id = NULL, videoProperties = NULL)
}
\arguments{
\item{url}{(string)
An URL to a video. The URL is valid as long as the source video exists and sharing settings do not change.}

\item{source}{(string)
The video source.
\itemize{
\item \code{SOURCE_UNSPECIFIED}: The video source is unspecified.
\item \code{YOUTUBE}: The video source is YouTube.
\item \code{DRIVE}: The video source is Google Drive.
}}

\item{id}{(string)
The video source's unique identifier for this video.}

\item{videoProperties}{(\link{VideoProperties})
The properties of the video.}
}
\value{
Video object
}
\description{
Video Object
}
\details{
A PageElement kind representing a video.
}
