% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{Video}
\alias{Video}
\title{Video Object}
\usage{
Video(source = NULL, videoProperties = NULL, id = NULL, url = NULL)
}
\arguments{
\item{source}{The video source}

\item{videoProperties}{The properties of the video}

\item{id}{The video source's unique identifier for this video}

\item{url}{An URL to a video}
}
\value{
Video object
}
\description{
Video Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A PageElement kind representing a video.
}
\concept{Video functions}
