% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{VideoProperties}
\alias{VideoProperties}
\title{VideoProperties Object}
\usage{
VideoProperties(
  start = NULL,
  end = NULL,
  outline = NULL,
  mute = NULL,
  autoPlay = NULL
)
}
\arguments{
\item{start}{The time at which to start playback, measured in seconds from the beginning of the video}

\item{end}{The time at which to end playback, measured in seconds from the beginning of the video}

\item{outline}{The outline of the video}

\item{mute}{Whether to mute the audio during video playback}

\item{autoPlay}{Whether to enable video autoplay when the page is displayed in present mode}
}
\value{
VideoProperties object
}
\description{
VideoProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of the Video.
}
\concept{VideoProperties functions}
