% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{WeightedFontFamily}
\alias{WeightedFontFamily}
\title{WeightedFontFamily Object}
\usage{
WeightedFontFamily(fontFamily = NULL, weight = NULL)
}
\arguments{
\item{fontFamily}{The font family of the text}

\item{weight}{The rendered weight of the text}
}
\value{
WeightedFontFamily object
}
\description{
WeightedFontFamily Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a font family and weight used to style a TextRun.
}
\concept{WeightedFontFamily functions}
