% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_objects.R
\name{WordArt}
\alias{WordArt}
\title{WordArt Object}
\usage{
WordArt(renderedText = NULL)
}
\arguments{
\item{renderedText}{(string)
The text rendered as word art.}
}
\value{
WordArt object
}
\description{
WordArt Object
}
\details{
A PageElement kind representing word art.
}
